refactor: extract SFTP file management
