+++ sshpilot/window.py
@@ -35,1390 +35,11 @@ from .key_manager import KeyManager, SSHKey
 # Port forwarding UI is now integrated into connection_dialog.py
 from .connection_dialog import ConnectionDialog
 from .askpass_utils import ensure_askpass_script
+from .preferences import PreferencesWindow, is_running_in_flatpak
+from .sshcopyid_window import SshCopyIdWindow
+from .sftp_utils import open_remote_in_file_manager
 
 logger = logging.getLogger(__name__)
-
-
-def is_running_in_flatpak() -> bool:
-    """Check if running inside Flatpak sandbox"""
-    return os.path.exists("/.flatpak-info") or os.environ.get("FLATPAK_ID") is not None
-
-def open_remote_in_file_manager(user: str, host: str, port: Optional[int] = None,
-                               path: Optional[str] = None, error_callback: Optional[Callable] = None,
-                               parent_window=None) -> Tuple[bool, Optional[str]]:
-    """Open remote server in file manager using SFTP URI with asynchronous verification"""
-
-    # Build sftp URI
-    p = path or "/"
-    port_part = f":{port}" if port else ""
-    uri = f"sftp://{user}@{host}{port_part}{p}"
-
-    logger.info(f"Opening SFTP URI: {uri}")
-
-    # Create progress dialog and start verification asynchronously
-    progress_dialog = MountProgressDialog(user, host, parent_window)
-    progress_dialog.present()
-    progress_dialog.start_progress_updates()
-    progress_dialog.update_progress(0.05, "Verifying SSH connection...")
-
-    def _on_verify_complete(success: bool):
-        if progress_dialog.is_cancelled:
-            return
-        if not success:
-            error_msg = "SSH connection failed - check credentials and network connectivity"
-            logger.error(f"SSH verification failed for {user}@{host}")
-            progress_dialog.update_progress(0.0, "SSH connection failed")
-            progress_dialog.show_error(error_msg)
-            GLib.timeout_add(1500, lambda: GLib.idle_add(progress_dialog.close))
-            if error_callback:
-                error_callback(error_msg)
-            return
-
-        logger.info(f"SSH connection verified for {user}@{host}")
-        progress_dialog.update_progress(0.3, "SSH verified, mounting...")
-        if is_running_in_flatpak():
-            _open_sftp_flatpak_compatible(uri, user, host, port, error_callback, progress_dialog)
-        else:
-            _mount_and_open_sftp(uri, user, host, error_callback, progress_dialog)
-
-    _verify_ssh_connection_async(user, host, port, _on_verify_complete)
-
-    return True, None
-
-def _mount_and_open_sftp(uri: str, user: str, host: str, error_callback=None, progress_dialog=None):
-    """Mount SFTP location and open in file manager"""
-    try:
-        logger.info(f"Mounting SFTP location: {uri}")
-
-        # Create progress dialog if not provided
-        if progress_dialog is None:
-            progress_dialog = _create_mount_progress_dialog(user, host)
-            progress_dialog.present()
-
-        gfile = Gio.File.new_for_uri(uri)
-        op = Gio.MountOperation()
-
-        def on_mounted(source, res, data=None):
-            try:
-                source.mount_enclosing_volume_finish(res)
-                logger.info(f"SFTP mount successful for {user}@{host}, opening file manager...")
-
-                # Update progress dialog
-                progress_dialog.update_progress(1.0, "Mount successful! Opening file manager...")
-
-                # Now it's mounted → open in the default file manager
-                Gio.AppInfo.launch_default_for_uri(uri, None)
-                logger.info(f"File manager launched successfully for {user}@{host}")
-
-                # Close progress dialog after a short delay
-                GLib.timeout_add(1000, lambda: GLib.idle_add(progress_dialog.close))
-
-            except GLib.Error as e:
-                # Check if the error is "already mounted" - this is actually a success case
-                if "already mounted" in e.message.lower():
-                    logger.info(f"SFTP location already mounted for {user}@{host}, opening file manager...")
-
-                    # Update progress dialog
-                    progress_dialog.update_progress(1.0, "Location already mounted! Opening file manager...")
-
-                    # Open in the default file manager
-                    Gio.AppInfo.launch_default_for_uri(uri, None)
-                    logger.info(f"File manager launched successfully for {user}@{host}")
-
-                    # Close progress dialog after a short delay
-                    GLib.timeout_add(1000, lambda: GLib.idle_add(progress_dialog.close))
-                else:
-                    error_msg = f"Could not mount {uri}: {e.message}"
-                    logger.error(f"Mount failed for {user}@{host}: {error_msg}")
-                    progress_dialog.update_progress(0.0, f"Mount failed: {e.message}")
-                    progress_dialog.show_error(error_msg)
-
-                    # Try Flatpak-compatible methods as fallback
-                    if is_running_in_flatpak():
-                        logger.info("Falling back to Flatpak-compatible methods")
-                        GLib.idle_add(progress_dialog.close)
-                        success, msg = _try_flatpak_compatible_mount(uri, user, host, None, error_callback)
-                        if not success and error_callback:
-                            error_callback(msg)
-                    else:
-                        GLib.timeout_add(1500, lambda: GLib.idle_add(progress_dialog.close))
-                        if error_callback:
-                            error_callback(error_msg)
-            except Exception as e:
-                error_msg = f"Unexpected error during mount: {str(e)}"
-                logger.error(f"Mount error for {user}@{host}: {e}")
-                progress_dialog.update_progress(0.0, f"Error: {str(e)}")
-                progress_dialog.show_error(error_msg)
-                GLib.timeout_add(1500, lambda: GLib.idle_add(progress_dialog.close))
-                if error_callback:
-                    error_callback(error_msg)
-
-        # Start progress updates if not already running
-        if not getattr(progress_dialog, 'progress_timer', None):
-            progress_dialog.start_progress_updates()
-
-        gfile.mount_enclosing_volume(
-            Gio.MountMountFlags.NONE,
-            op,
-            None,  # cancellable
-            on_mounted,
-            None
-        )
-
-        logger.info(f"Mount operation started for {user}@{host}")
-        return True, None
-        
-    except Exception as e:
-        error_msg = f"Failed to start mount operation: {str(e)}"
-        logger.error(f"Mount operation failed for {user}@{host}: {e}")
-        GLib.timeout_add(1500, lambda: GLib.idle_add(progress_dialog.close))
-
-        # Try Flatpak-compatible methods as fallback
-        if is_running_in_flatpak():
-            logger.info("Primary mount failed, trying Flatpak-compatible methods")
-            return _try_flatpak_compatible_mount(uri, user, host, None, error_callback)
-
-        if error_callback:
-            error_callback(error_msg)
-        return False, error_msg
-
-def _verify_ssh_connection(user: str, host: str, port: Optional[int]) -> bool:
-    """Verify SSH connection without full mount"""
-    ssh_cmd = ["ssh", "-o", "ConnectTimeout=10", "-o", "BatchMode=yes", 
-               "-o", "StrictHostKeyChecking=accept-new"]
-    if port:
-        ssh_cmd.extend(["-p", str(port)])
-    ssh_cmd.extend([f"{user}@{host}", "echo", "READY"])
-    
-    try:
-        result = subprocess.run(ssh_cmd, capture_output=True, text=True, timeout=15)
-        return result.returncode == 0
-    except (subprocess.TimeoutExpired, Exception):
-        return False
-
-def _verify_ssh_connection_async(user: str, host: str, port: Optional[int], callback: Callable[[bool], None]) -> None:
-    """Verify SSH connection on a background thread and invoke callback with the result"""
-
-    def worker():
-        success = _verify_ssh_connection(user, host, port)
-        GLib.idle_add(callback, success)
-
-    threading.Thread(target=worker, daemon=True).start()
-
-def _open_sftp_flatpak_compatible(uri: str, user: str, host: str, port: Optional[int],
-                                 error_callback: Optional[Callable], progress_dialog=None,
-                                 parent_window=None) -> Tuple[bool, Optional[str]]:
-    """Open SFTP using Flatpak-compatible methods with proper portal usage"""
-
-    # Reuse existing progress dialog if provided
-    if progress_dialog is None:
-        progress_dialog = MountProgressDialog(user, host, parent_window)
-        progress_dialog.present()
-    if not getattr(progress_dialog, 'progress_timer', None):
-        progress_dialog.start_progress_updates()
-    
-    # Method 1: Use XDG Desktop Portal File Chooser to access GVFS mounts
-    try:
-        progress_dialog.update_progress(0.2, "Trying portal file access...")
-        success = _try_portal_file_access(uri, user, host)
-        if success:
-            progress_dialog.update_progress(1.0, "Success! Opening file manager...")
-            GLib.timeout_add(1000, lambda: GLib.idle_add(progress_dialog.close))
-            return True, None
-    except Exception as e:
-        logger.warning(f"Portal file access failed: {e}")
-    
-    # Method 2: Try to launch external file manager that can handle SFTP
-    try:
-        progress_dialog.update_progress(0.4, "Trying external file managers...")
-        success = _try_external_file_managers(uri, user, host)
-        if success:
-            progress_dialog.update_progress(1.0, "Success! File manager opened...")
-            GLib.timeout_add(1000, lambda: GLib.idle_add(progress_dialog.close))
-            return True, None
-    except Exception as e:
-        logger.warning(f"External file managers failed: {e}")
-    
-    # Method 3: Use host's GVFS if accessible
-    try:
-        progress_dialog.update_progress(0.6, "Checking host GVFS mounts...")
-        success = _try_host_gvfs_access(uri, user, host, port)
-        if success:
-            progress_dialog.update_progress(1.0, "Success! Found existing mount...")
-            GLib.timeout_add(1000, lambda: GLib.idle_add(progress_dialog.close))
-            return True, None
-    except Exception as e:
-        logger.warning(f"Host GVFS access failed: {e}")
-    
-    # Method 4: Show connection dialog for manual setup
-    progress_dialog.update_progress(0.8, "Preparing manual connection options...")
-    success = _show_manual_connection_dialog(user, host, port, uri)
-    if success:
-        GLib.idle_add(progress_dialog.close)
-        return True, "Manual connection dialog opened"
-    
-    # All methods failed
-    error_msg = "Could not open SFTP connection - try mounting the location manually first"
-    progress_dialog.show_error(error_msg)
-    GLib.timeout_add(1500, lambda: GLib.idle_add(progress_dialog.close))
-    if error_callback:
-        GLib.idle_add(error_callback, error_msg)
-    return False, error_msg
-
-def _try_portal_file_access(uri: str, user: str, host: str) -> bool:
-    """Try to access SFTP location via XDG Desktop Portal - skip file chooser"""
-    
-    # Skip the file chooser dialog approach - it's not what we want
-    # Instead, try to use the portal to trigger a mount and then open directly
-    try:
-        # Try to mount via D-Bus portal interface directly
-        return _try_dbus_gvfs_mount(uri, user, host)
-    except Exception as e:
-        logger.warning(f"Portal D-Bus mount failed: {e}")
-        return False
-
-def _try_dbus_gvfs_mount(uri: str, user: str, host: str) -> bool:
-    """Try to mount GVFS via gio command and open directly"""
-    
-    try:
-        if subprocess.run(["which", "flatpak-spawn"], capture_output=True).returncode == 0:
-            # Check if gio is available on host (it should be)
-            check_cmd = ["flatpak-spawn", "--host", "which", "gio"]
-            if subprocess.run(check_cmd, capture_output=True).returncode == 0:
-                logger.info(f"Using gio mount for {uri}")
-                
-                # Mount using host's gio mount
-                mount_cmd = ["flatpak-spawn", "--host", "gio", "mount", uri]
-                result = subprocess.run(mount_cmd, capture_output=True, text=True, timeout=30)
-                
-                logger.info(f"gio mount result: returncode={result.returncode}, stdout={result.stdout}, stderr={result.stderr}")
-                
-                if result.returncode == 0 or "already mounted" in result.stderr.lower() or "Operation not supported" not in result.stderr:
-                    logger.info(f"gio mount successful or location already accessible")
-                    
-                    # Give it a moment for the mount to be ready
-                    import time
-                    time.sleep(1)
-                    
-                    # Find the actual mount point and open that instead of the URI
-                    mount_point = _find_gvfs_mount_point(user, host)
-                    if mount_point:
-                        open_cmd = ["flatpak-spawn", "--host", "xdg-open", mount_point]
-                        subprocess.Popen(open_cmd, stdout=subprocess.DEVNULL, stderr=subprocess.DEVNULL)
-                        logger.info(f"File manager opened at mount point: {mount_point}")
-                        return True
-                    else:
-                        # Try opening the URI directly with specific file managers
-                        logger.info("Mount point not found, trying file managers with URI")
-                        return _try_specific_file_managers_with_uri(uri)
-                else:
-                    logger.warning(f"gio mount failed: {result.stderr}")
-                    # Still try file managers in case they can handle it
-                    return _try_specific_file_managers_with_uri(uri)
-            else:
-                logger.warning("gio not available on host system")
-                return _try_specific_file_managers_with_uri(uri)
-        
-        return False
-        
-    except Exception as e:
-        logger.warning(f"gio mount failed: {e}")
-        return False
-
-def _find_gvfs_mount_point(user: str, host: str) -> Optional[str]:
-    """Find the actual GVFS mount point for the SFTP connection"""
-    
-    gvfs_paths = [
-        f"/run/user/{os.getuid()}/gvfs",
-        f"/var/run/user/{os.getuid()}/gvfs"
-    ]
-    
-    for gvfs_path in gvfs_paths:
-        try:
-            if os.path.exists(gvfs_path):
-                for mount_dir in os.listdir(gvfs_path):
-                    # Look for SFTP mount matching our host
-                    if f"sftp:host={host}" in mount_dir and f"user={user}" in mount_dir:
-                        mount_point = os.path.join(gvfs_path, mount_dir)
-                        logger.info(f"Found GVFS mount point: {mount_point}")
-                        return mount_point
-        except Exception as e:
-            logger.debug(f"Could not check GVFS path {gvfs_path}: {e}")
-    
-    return None
-
-def _try_specific_file_managers_with_uri(uri: str) -> bool:
-    """Try specific file managers that handle SFTP URIs properly"""
-    
-    # File managers that are known to handle SFTP URIs well
-    managers = [
-        ["flatpak-spawn", "--host", "nautilus", uri],
-        ["flatpak-spawn", "--host", "thunar", uri],
-        ["flatpak-spawn", "--host", "dolphin", uri],
-        ["flatpak-spawn", "--host", "nemo", uri]
-    ]
-    
-    for cmd in managers:
-        try:
-            # Check if the file manager exists on host
-            check_cmd = ["flatpak-spawn", "--host", "which", cmd[2]]
-            if subprocess.run(check_cmd, capture_output=True).returncode == 0:
-                logger.info(f"Trying {cmd[2]} with SFTP URI")
-                subprocess.Popen(cmd, stdout=subprocess.DEVNULL, stderr=subprocess.DEVNULL)
-                logger.info(f"Launched {cmd[2]} via flatpak-spawn")
-                return True
-        except Exception as e:
-            logger.debug(f"Failed to launch {cmd[2]}: {e}")
-    
-    return False
-
-def _try_external_file_managers(uri: str, user: str, host: str) -> bool:
-    """Try launching external file managers that handle SFTP"""
-    
-    # File managers with their specific SFTP handling
-    managers = [
-        ("nautilus", [uri]),  # Usually handles SFTP URIs well
-        ("thunar", [uri]),    # Good SFTP support
-        ("dolphin", [uri]),   # KDE file manager
-        ("nemo", [uri]),      # Cinnamon file manager
-        ("pcmanfm", [uri]),   # Lightweight option
-    ]
-    
-    for manager, cmd in managers:
-        try:
-            # Use flatpak-spawn to run on the host if available
-            if subprocess.run(["which", "flatpak-spawn"], capture_output=True).returncode == 0:
-                # Check if manager exists on host
-                check_cmd = ["flatpak-spawn", "--host", "which", manager]
-                if subprocess.run(check_cmd, capture_output=True).returncode == 0:
-                    spawn_cmd = ["flatpak-spawn", "--host", manager, uri]
-                    subprocess.Popen(spawn_cmd, stdout=subprocess.DEVNULL, stderr=subprocess.DEVNULL)
-                    logger.info(f"Launched {manager} via flatpak-spawn with URI")
-                    return True
-            
-            # Try direct launch as fallback
-            elif subprocess.run(["which", manager], capture_output=True).returncode == 0:
-                subprocess.Popen(cmd, stdout=subprocess.DEVNULL, stderr=subprocess.DEVNULL)
-                logger.info(f"Launched {manager} directly with URI")
-                return True
-                
-        except Exception as e:
-            logger.debug(f"Failed to launch {manager}: {e}")
-            continue
-    
-    # If no file manager worked with URI, try a different approach
-    logger.warning("No file manager could handle SFTP URI directly")
-    return _try_alternative_approaches(uri, user, host)
-
-def _try_alternative_approaches(uri: str, user: str, host: str) -> bool:
-    """Try alternative approaches when direct URI opening fails"""
-    
-    # Method 1: Try opening network locations in file managers
-    network_locations = [
-        "network:///",
-        "sftp://",
-        f"sftp://{host}/"
-    ]
-    
-    for location in network_locations:
-        try:
-            if subprocess.run(["which", "flatpak-spawn"], capture_output=True).returncode == 0:
-                cmd = ["flatpak-spawn", "--host", "nautilus", location]
-                subprocess.Popen(cmd, stdout=subprocess.DEVNULL, stderr=subprocess.DEVNULL)
-                logger.info(f"Opened network location: {location}")
-                return True
-        except Exception as e:
-            logger.debug(f"Failed to open {location}: {e}")
-    
-    # Method 2: Try using gio mount command
-    try:
-        if subprocess.run(["which", "flatpak-spawn"], capture_output=True).returncode == 0:
-            # Check if gio is available
-            check_cmd = ["flatpak-spawn", "--host", "which", "gio"]
-            if subprocess.run(check_cmd, capture_output=True).returncode == 0:
-                # Use gio to mount
-                mount_cmd = ["flatpak-spawn", "--host", "gio", "mount", uri]
-                result = subprocess.run(mount_cmd, capture_output=True, text=True, timeout=30)
-                
-                if result.returncode == 0:
-                    # Try to open with nautilus after mounting
-                    open_cmd = ["flatpak-spawn", "--host", "nautilus", uri]
-                    subprocess.Popen(open_cmd, stdout=subprocess.DEVNULL, stderr=subprocess.DEVNULL)
-                    logger.info(f"Mounted with gio and opened with nautilus")
-                    return True
-    except Exception as e:
-        logger.debug(f"gio mount failed: {e}")
-    
-    return False
-
-def _try_host_gvfs_access(uri: str, user: str, host: str, port: Optional[int]) -> bool:
-    """Try accessing GVFS mounts on the host system"""
-    
-    # Check if we can access the host's GVFS mounts
-    gvfs_paths = [
-        f"/run/user/{os.getuid()}/gvfs",
-        f"/var/run/user/{os.getuid()}/gvfs",
-        f"{os.path.expanduser('~')}/.gvfs"
-    ]
-    
-    for gvfs_path in gvfs_paths:
-        if os.path.exists(gvfs_path):
-            # Look for existing SFTP mount
-            sftp_pattern = f"sftp:host={host}"
-            try:
-                for mount_dir in os.listdir(gvfs_path):
-                    if sftp_pattern in mount_dir:
-                        mount_path = os.path.join(gvfs_path, mount_dir)
-                        # Open in file manager
-                        subprocess.Popen(["xdg-open", mount_path], 
-                                       stdout=subprocess.DEVNULL, stderr=subprocess.DEVNULL)
-                        logger.info(f"Opened existing GVFS mount: {mount_path}")
-                        return True
-            except Exception as e:
-                logger.debug(f"Could not access GVFS path {gvfs_path}: {e}")
-                continue
-    
-    return False
-
-def _show_manual_connection_dialog(user: str, host: str, port: Optional[int], uri: str) -> bool:
-    """Show dialog with manual connection instructions"""
-    
-    dialog = SftpConnectionDialog(user, host, port, uri)
-    dialog.present()
-    return True
-
-def _open_sftp_native(uri: str, user: str, host: str, error_callback: Optional[Callable], parent_window=None) -> Tuple[bool, Optional[str]]:
-    """Native installation SFTP opening with GVFS"""
-    
-    # Try direct GVFS mount first
-    try:
-        success = _mount_and_open_sftp_native(uri, user, host, error_callback, parent_window)
-        if success:
-            return True, None
-    except Exception as e:
-        logger.warning(f"Native GVFS mount failed: {e}")
-    
-    # Fall back to Flatpak-compatible methods
-    return _open_sftp_flatpak_compatible(uri, user, host, None, error_callback, parent_window=parent_window)
-
-def _mount_and_open_sftp_native(uri: str, user: str, host: str, error_callback: Optional[Callable], parent_window=None) -> bool:
-    """Original native GVFS mounting method"""
-    
-    logger.info(f"Mounting SFTP location: {uri}")
-    
-    progress_dialog = MountProgressDialog(user, host, parent_window)
-    progress_dialog.present()
-    
-    gfile = Gio.File.new_for_uri(uri)
-    op = Gio.MountOperation()
-    
-    def on_mounted(source, res, data=None):
-        try:
-            source.mount_enclosing_volume_finish(res)
-            logger.info(f"SFTP mount successful for {user}@{host}")
-            progress_dialog.update_progress(1.0, "Mount successful! Opening file manager...")
-            Gio.AppInfo.launch_default_for_uri(uri, None)
-            GLib.timeout_add(1000, progress_dialog.close)
-            
-        except GLib.Error as e:
-            if "already mounted" in e.message.lower():
-                logger.info(f"SFTP location already mounted for {user}@{host}")
-                progress_dialog.update_progress(1.0, "Location already mounted! Opening file manager...")
-                Gio.AppInfo.launch_default_for_uri(uri, None)
-                GLib.timeout_add(1000, progress_dialog.close)
-            else:
-                error_msg = f"Could not mount {uri}: {e.message}"
-                logger.error(f"Mount failed for {user}@{host}: {error_msg}")
-                progress_dialog.show_error(error_msg)
-                if error_callback:
-                    error_callback(error_msg)
-    
-    progress_dialog.start_progress_updates()
-    gfile.mount_enclosing_volume(Gio.MountMountFlags.NONE, op, None, on_mounted, None)
-    logger.info(f"Mount operation started for {user}@{host}")
-    return True
-
-def _try_flatpak_compatible_mount(uri: str, user: str, host: str, port: int|None, error_callback=None):
-    """Try various methods to open SFTP in Flatpak environment"""
-    
-    # Method 1: Try direct URI launch (sometimes works if gvfs is available)
-    try:
-        logger.info("Trying direct URI launch...")
-        Gio.AppInfo.launch_default_for_uri(uri, None)
-        logger.info(f"Direct URI launch successful for {user}@{host}")
-        return True, None
-    except Exception as e:
-        logger.warning(f"Direct URI launch failed: {e}")
-    
-    # Method 2: Try external file managers that handle SFTP
-    external_managers = [
-        ("nautilus", [uri]),  # GNOME Files
-        ("thunar", [uri]),    # XFCE Thunar
-        ("dolphin", [uri]),   # KDE Dolphin
-        ("pcmanfm", [uri]),   # PCManFM
-        ("nemo", [uri]),      # Nemo
-    ]
-    
-    for manager, cmd in external_managers:
-        try:
-            # Check if the file manager exists
-            if subprocess.run(["which", manager], capture_output=True).returncode == 0:
-                logger.info(f"Trying external file manager: {manager}")
-                subprocess.Popen(cmd, stdout=subprocess.DEVNULL, stderr=subprocess.DEVNULL)
-                logger.info(f"External file manager {manager} launched for {user}@{host}")
-                return True, None
-        except Exception as e:
-            logger.warning(f"Failed to launch {manager}: {e}")
-    
-    # Method 3: Try mounting via command line tools
-    return _try_command_line_mount(user, host, port, error_callback)
-
-def _try_command_line_mount(user: str, host: str, port: int|None, error_callback=None):
-    """Try mounting via command line utilities"""
-    
-    # Create a mount point in user's home directory
-    mount_point = os.path.expanduser(f"~/sftp-{host}")
-    
-    try:
-        # Create mount point if it doesn't exist
-        os.makedirs(mount_point, exist_ok=True)
-        
-        # Try sshfs if available
-        sshfs_cmd = ["sshfs"]
-        if port:
-            sshfs_cmd.extend(["-p", str(port)])
-        
-        sshfs_cmd.extend([
-            "-o", "reconnect,ServerAliveInterval=15,ServerAliveCountMax=3",
-            f"{user}@{host}:/",
-            mount_point
-        ])
-        
-        logger.info(f"Trying sshfs mount: {' '.join(sshfs_cmd)}")
-        result = subprocess.run(sshfs_cmd, capture_output=True, text=True, timeout=30)
-        
-        if result.returncode == 0:
-            logger.info(f"sshfs mount successful at {mount_point}")
-            # Open the mount point in file manager
-            try:
-                subprocess.Popen(["xdg-open", mount_point])
-                return True, None
-            except Exception as e:
-                logger.warning(f"Failed to open mount point: {e}")
-                return True, f"Mounted at {mount_point} but couldn't open file manager"
-        else:
-            logger.warning(f"sshfs failed: {result.stderr}")
-            
-    except subprocess.TimeoutExpired:
-        logger.error("sshfs mount timeout")
-    except Exception as e:
-        logger.error(f"sshfs mount error: {e}")
-    
-    # Method 4: Fall back to terminal-based SFTP client
-    return _launch_terminal_sftp(user, host, port, error_callback)
-
-def _launch_terminal_sftp(user: str, host: str, port: int|None, error_callback=None):
-    """Launch terminal-based SFTP client as last resort"""
-    
-    terminals = ["gnome-terminal", "konsole", "xfce4-terminal", "xterm"]
-    
-    sftp_cmd = ["sftp"]
-    if port:
-        sftp_cmd.extend(["-P", str(port)])
-    sftp_cmd.append(f"{user}@{host}")
-    
-    for terminal in terminals:
-        try:
-            if subprocess.run(["which", terminal], capture_output=True).returncode == 0:
-                logger.info(f"Launching terminal SFTP with {terminal}")
-                
-                if terminal == "gnome-terminal":
-                    cmd = [terminal, "--", *sftp_cmd]
-                elif terminal == "konsole":
-                    cmd = [terminal, "-e", *sftp_cmd]
-                elif terminal == "xfce4-terminal":
-                    cmd = [terminal, "-e", " ".join(sftp_cmd)]
-                else:
-                    cmd = [terminal, "-e", " ".join(sftp_cmd)]
-                
-                subprocess.Popen(cmd, stdout=subprocess.DEVNULL, stderr=subprocess.DEVNULL)
-                return True, f"Opened SFTP connection in {terminal}"
-                
-        except Exception as e:
-            logger.warning(f"Failed to launch {terminal}: {e}")
-    
-    error_msg = "Could not open SFTP connection - no compatible file manager or terminal found"
-    logger.error(error_msg)
-    if error_callback:
-        error_callback(error_msg)
-        return False, error_msg
-
-def _create_mount_progress_dialog(user: str, host: str):
-    """Create a progress dialog for SFTP mount operation"""
-    return MountProgressDialog(user, host)
-
-class MountProgressDialog(Adw.Window):
-    """Progress dialog for SFTP mount operations"""
-    
-    def __init__(self, user: str, host: str, parent_window=None):
-        super().__init__()
-        self.user = user
-        self.host = host
-        self.progress_value = 0.0
-        self.is_cancelled = False
-        self.progress_timer = None
-        
-        self.set_title("Mounting SFTP Connection")
-        self.set_default_size(500, 200)
-        self.set_resizable(False)
-        self.set_modal(True)
-        
-        # Set as transient for parent window if provided
-        if parent_window:
-            self.set_transient_for(parent_window)
-        
-        # Main container
-        main_box = Gtk.Box(orientation=Gtk.Orientation.VERTICAL, spacing=12)
-        main_box.set_margin_start(20)
-        main_box.set_margin_end(20)
-        main_box.set_margin_top(20)
-        main_box.set_margin_bottom(20)
-        
-        # Header
-        header_label = Gtk.Label()
-        header_label.set_markup(f"<b>Connecting to {user}@{host}</b>")
-        main_box.append(header_label)
-        
-        # Progress bar
-        self.progress_bar = Gtk.ProgressBar()
-        self.progress_bar.set_show_text(True)
-        self.progress_bar.set_text("Initializing connection...")
-        main_box.append(self.progress_bar)
-        
-        # Status label
-        self.status_label = Gtk.Label()
-        self.status_label.set_text("Preparing SFTP mount...")
-        main_box.append(self.status_label)
-        
-        # Cancel button
-        self.cancel_button = Gtk.Button.new_with_label("Cancel")
-        self.cancel_button.connect('clicked', self._on_cancel)
-        self.cancel_button.set_halign(Gtk.Align.CENTER)
-        main_box.append(self.cancel_button)
-        
-        self.set_content(main_box)
-    
-    def _on_cancel(self, button):
-        """Cancel mount operation"""
-        self.is_cancelled = True
-        if self.progress_timer:
-            GLib.source_remove(self.progress_timer)
-        self.close()
-    
-    def start_progress_updates(self):
-        """Start simulated progress updates"""
-        self.progress_value = 0.0
-        self.progress_timer = GLib.timeout_add(100, self._update_progress_simulation)
-        
-    def _update_progress_simulation(self):
-        """Simulate mounting progress"""
-        if self.is_cancelled:
-            return False
-            
-        self.progress_value += 0.02
-        if self.progress_value >= 0.9:
-            self.progress_value = 0.9
-            
-        self.update_progress(self.progress_value, "Establishing SFTP connection...")
-        return True
-        
-    def update_progress(self, fraction: float, text: str):
-        """Update progress bar and status"""
-        self.progress_bar.set_fraction(fraction)
-        self.progress_bar.set_text(text)
-        self.status_label.set_text(text)
-        
-    def show_error(self, error_text: str):
-        """Show error state"""
-        self.progress_bar.add_css_class("error")
-        self.cancel_button.set_label("Close")
-    
-    def close(self, widget=None):
-        """Close the dialog"""
-        if self.progress_timer:
-            GLib.source_remove(self.progress_timer)
-        self.destroy()
-
-class SftpConnectionDialog(Adw.Window):
-    """Dialog showing manual connection options for SFTP"""
-    
-    def __init__(self, user: str, host: str, port: Optional[int], uri: str):
-        super().__init__()
-        self.user = user
-        self.host = host  
-        self.port = port
-        self.uri = uri
-        
-        self.set_title("SFTP Connection")
-        self.set_default_size(600, 400)
-        self.set_modal(True)
-        
-        # Main container
-        main_box = Gtk.Box(orientation=Gtk.Orientation.VERTICAL, spacing=16)
-        main_box.set_margin_start(24)
-        main_box.set_margin_end(24)
-        main_box.set_margin_top(24)
-        main_box.set_margin_bottom(24)
-        
-        # Header
-        header_box = Gtk.Box(orientation=Gtk.Orientation.HORIZONTAL, spacing=16)
-        
-        icon = Gtk.Image.new_from_icon_name("folder-remote-symbolic")
-        icon.set_pixel_size(48)
-        header_box.append(icon)
-        
-        title_box = Gtk.Box(orientation=Gtk.Orientation.VERTICAL, spacing=4)
-        title_label = Gtk.Label()
-        title_label.set_markup("<b>SFTP Connection Options</b>")
-        title_label.set_halign(Gtk.Align.START)
-        title_box.append(title_label)
-        
-        subtitle_label = Gtk.Label()
-        subtitle_label.set_text(f"Connect to {user}@{host}")
-        subtitle_label.add_css_class("dim-label")
-        subtitle_label.set_halign(Gtk.Align.START)
-        title_box.append(subtitle_label)
-        
-        header_box.append(title_box)
-        main_box.append(header_box)
-        
-        # Instructions
-        instructions = Gtk.Label()
-        instructions.set_markup(
-            "Due to Flatpak security restrictions, direct SFTP mounting is limited.\n"
-            "Please choose one of the following options:"
-        )
-        instructions.set_wrap(True)
-        instructions.set_halign(Gtk.Align.START)
-        main_box.append(instructions)
-        
-        # Options list
-        options_box = Gtk.Box(orientation=Gtk.Orientation.VERTICAL, spacing=12)
-        
-        # Option 1: File Manager
-        option1_box = self._create_option_box(
-            "Open in File Manager",
-            "Try opening the SFTP location directly in your system's file manager",
-            "folder-symbolic",
-            lambda: self._try_file_manager()
-        )
-        options_box.append(option1_box)
-        
-        # Option 2: Copy URI
-        option2_box = self._create_option_box(
-            "Copy SFTP URI",
-            f"Copy the SFTP URI to clipboard: {uri}",
-            "edit-copy-symbolic", 
-            lambda: self._copy_uri()
-        )
-        options_box.append(option2_box)
-        
-        # Option 3: Terminal
-        option3_box = self._create_option_box(
-            "Open in Terminal",
-            "Open an SFTP connection in terminal",
-            "utilities-terminal-symbolic",
-            lambda: self._open_terminal()
-        )
-        options_box.append(option3_box)
-        
-        main_box.append(options_box)
-        
-        # Close button
-        close_button = Gtk.Button.new_with_label("Close")
-        close_button.connect('clicked', lambda b: self.close())
-        close_button.set_halign(Gtk.Align.END)
-        main_box.append(close_button)
-        
-        self.set_content(main_box)
-        
-    def _create_option_box(self, title: str, description: str, icon_name: str, callback: Callable) -> Gtk.Box:
-        """Create an option box with icon, text, and button"""
-        
-        box = Gtk.Box(orientation=Gtk.Orientation.HORIZONTAL, spacing=12)
-        box.add_css_class("card")
-        box.set_margin_start(12)
-        box.set_margin_end(12) 
-        box.set_margin_top(8)
-        box.set_margin_bottom(8)
-        
-        icon = Gtk.Image.new_from_icon_name(icon_name)
-        icon.set_pixel_size(24)
-        box.append(icon)
-        
-        text_box = Gtk.Box(orientation=Gtk.Orientation.VERTICAL, spacing=2)
-        text_box.set_hexpand(True)
-        
-        title_label = Gtk.Label()
-        title_label.set_markup(f"<b>{title}</b>")
-        title_label.set_halign(Gtk.Align.START)
-        text_box.append(title_label)
-        
-        desc_label = Gtk.Label()
-        desc_label.set_text(description)
-        desc_label.set_halign(Gtk.Align.START)
-        desc_label.add_css_class("dim-label")
-        desc_label.set_wrap(True)
-        text_box.append(desc_label)
-        
-        box.append(text_box)
-        
-        button = Gtk.Button.new_with_label("Try")
-        button.connect('clicked', lambda b: callback())
-        box.append(button)
-        
-        return box
-    
-    def _try_file_manager(self):
-        """Try opening in file manager"""
-        try:
-            Gio.AppInfo.launch_default_for_uri(self.uri, None)
-        except Exception:
-            # Fall back to xdg-open
-            subprocess.Popen(["xdg-open", self.uri], 
-                           stdout=subprocess.DEVNULL, stderr=subprocess.DEVNULL)
-    
-    def _copy_uri(self):
-        """Copy URI to clipboard"""
-        clipboard = Gtk.Clipboard.get(Gdk.SELECTION_CLIPBOARD)
-        clipboard.set_text(self.uri, -1)
-    
-    def _open_terminal(self):
-        """Open SFTP in terminal"""
-        sftp_cmd = ["sftp"]
-        if self.port:
-            sftp_cmd.extend(["-P", str(self.port)])
-        sftp_cmd.append(f"{self.user}@{self.host}")
-        
-        terminals = ["gnome-terminal", "konsole", "xfce4-terminal", "xterm"]
-        
-        for terminal in terminals:
-            try:
-                if subprocess.run(["which", terminal], capture_output=True).returncode == 0:
-                    if terminal == "gnome-terminal":
-                        cmd = [terminal, "--", *sftp_cmd]
-                    elif terminal == "konsole":
-                        cmd = [terminal, "-e", *sftp_cmd] 
-                    else:
-                        cmd = [terminal, "-e", " ".join(sftp_cmd)]
-                    
-                    subprocess.Popen(cmd, stdout=subprocess.DEVNULL, stderr=subprocess.DEVNULL)
-                    break
-            except Exception:
-                continue
-
-
-# =========================
-# SSH Copy-ID Full Window
-# =========================
-class SshCopyIdWindow(Gtk.Window):
-    """
-    Full Adwaita-styled window for installing a public key on a server.
-    - Shows selected server nickname
-    - Two modes:
-        1) Use existing key (DropDown)
-        2) Generate new key (embedded key-generator form)
-    - Pressing OK triggers:
-        - Either copy selected existing key
-        - Or generate a new key, then copy it
-    - Uses your existing terminal flow:
-        parent._show_ssh_copy_id_terminal_using_main_widget(connection, ssh_key)
-    """
-
-    def __init__(self, parent, connection, key_manager, connection_manager):
-        logger.info("SshCopyIdWindow: Initializing window")
-        logger.debug(f"SshCopyIdWindow: Constructor called with connection: {getattr(connection, 'nickname', 'unknown')}")
-        logger.debug(f"SshCopyIdWindow: Connection object type: {type(connection)}")
-        logger.debug(f"SshCopyIdWindow: Key manager type: {type(key_manager)}")
-        logger.debug(f"SshCopyIdWindow: Connection manager type: {type(connection_manager)}")
-        
-        try:
-            super().__init__()
-            self.set_transient_for(parent)
-            self.set_modal(True)
-            self.set_title("Install Public Key on Server")
-            self.set_resizable(False)
-            self.set_default_size(500, 400)
-            logger.debug("SshCopyIdWindow: Base window initialized")
-
-            self._parent = parent
-            self._conn = connection
-            self._km = key_manager
-            self._cm = connection_manager
-            logger.debug("SshCopyIdWindow: Instance variables set")
-            
-            logger.info(f"SshCopyIdWindow: Window initialized for connection {getattr(connection, 'nickname', 'unknown')}")
-        except Exception as e:
-            logger.error(f"SshCopyIdWindow: Failed to initialize window: {e}")
-            logger.debug(f"SshCopyIdWindow: Exception details: {type(e).__name__}: {str(e)}")
-            raise
-
-        # ---------- Outer layout ----------
-        logger.info("SshCopyIdWindow: Creating outer layout")
-        try:
-            tv = Adw.ToolbarView()
-            self.set_child(tv)
-            
-            # ---------- Header Bar ----------
-            logger.info("SshCopyIdWindow: Creating header bar")
-            hb = Adw.HeaderBar()
-            tv.add_top_bar(hb)
-
-            # Cancel button
-            btn_cancel = Gtk.Button(label="Cancel")
-            btn_cancel.connect("clicked", self._on_close_clicked)
-            hb.pack_start(btn_cancel)
-
-            self.btn_ok = Gtk.Button(label="OK")
-            self.btn_ok.add_css_class("suggested-action")
-            self.btn_ok.connect("clicked", self._on_ok_clicked)
-            hb.pack_end(self.btn_ok)
-            logger.info("SshCopyIdWindow: Header bar created successfully")
-
-            content = Gtk.Box(orientation=Gtk.Orientation.VERTICAL, spacing=12)
-            content.set_margin_top(18); content.set_margin_bottom(18)
-            content.set_margin_start(18); content.set_margin_end(18)
-            tv.set_content(content)
-
-            # ---------- Intro text ----------
-            server_name = getattr(self._conn, "nickname", None) or \
-                          f"{getattr(self._conn, 'username', 'user')}@{getattr(self._conn, 'host', 'host')}"
-            
-            # Create a simple label instead of StatusPage for normal font size
-            intro_label = Gtk.Label()
-            intro_label.set_markup(f'Copy your public key to "{server_name}".')
-            intro_label.set_halign(Gtk.Align.CENTER)
-            intro_label.set_margin_bottom(12)
-            content.append(intro_label)
-            logger.info(f"SshCopyIdWindow: Intro text created for server: {server_name}")
-        except Exception as e:
-            logger.error(f"SshCopyIdWindow: Failed to create outer layout: {e}")
-            raise
-
-        # ---------- Options group ----------
-        logger.info("SshCopyIdWindow: Creating options group")
-        try:
-            group = Adw.PreferencesGroup(title="")
-
-            # Radio option 1: Use existing key (using CheckButton with group for radio behavior)
-            self.radio_existing = Gtk.CheckButton(label="Copy existing key")
-            self.radio_existing.set_can_focus(True)  # Make it focusable for tab navigation
-            self.radio_generate = Gtk.CheckButton(label="Generate new key")
-            self.radio_generate.set_can_focus(True)  # Make it focusable for tab navigation
-
-            # Make them behave like radio buttons (GTK4)
-            self.radio_generate.set_group(self.radio_existing)
-            self.radio_existing.set_active(True)
-            logger.info("SshCopyIdWindow: Radio buttons created successfully")
-        except Exception as e:
-            logger.error(f"SshCopyIdWindow: Failed to create radio buttons: {e}")
-            raise
-
-        # Existing key row with dropdown
-        logger.info("SshCopyIdWindow: Creating existing key dropdown")
-        try:
-            existing_box = Gtk.Box(orientation=Gtk.Orientation.HORIZONTAL, spacing=12)
-            existing_box.set_margin_start(12)
-            existing_box.set_margin_bottom(6)
-            self.dropdown_existing = Gtk.DropDown()
-            self.dropdown_existing.set_can_focus(True)  # Make it focusable for tab navigation
-            existing_box.append(Gtk.Label(label="Select key:", xalign=0))
-            existing_box.append(self.dropdown_existing)
-
-            # Fill dropdown with discovered keys
-            self._reload_existing_keys()
-            logger.info("SshCopyIdWindow: Existing key dropdown created successfully")
-        except Exception as e:
-            logger.error(f"SshCopyIdWindow: Failed to create existing key dropdown: {e}")
-            raise
-
-        # Generate form (embedded)
-        logger.info("SshCopyIdWindow: Creating key generation form")
-        try:
-            self.generate_revealer = Gtk.Revealer()
-            self.generate_revealer.set_reveal_child(False)
-            self.generate_revealer.set_transition_type(Gtk.RevealerTransitionType.SLIDE_DOWN)
-
-            gen_box = Gtk.Box(orientation=Gtk.Orientation.VERTICAL, spacing=8)
-            gen_box.set_margin_start(12)
-            gen_box.set_margin_top(6)
-            gen_box.set_can_focus(True)  # Make the box focusable for tab navigation
-
-            # Key name
-            self.row_key_name = Adw.EntryRow()
-            self.row_key_name.set_title("Key file name")
-            self.row_key_name.set_text("id_ed25519")
-            self.row_key_name.set_can_focus(True)  # Make it focusable for tab navigation
-            gen_box.append(self.row_key_name)
-
-            # Key type
-            key_type_box = Gtk.Box(orientation=Gtk.Orientation.HORIZONTAL, spacing=12)
-            key_type_label = Gtk.Label(label="Key type:", xalign=0)
-            self.type_dropdown = Gtk.DropDown()
-            self._types_model = Gtk.StringList.new(["ed25519", "rsa"])
-            self.type_dropdown.set_model(self._types_model)
-            self.type_dropdown.set_selected(0)
-            self.type_dropdown.set_can_focus(True)  # Make it focusable for tab navigation
-            key_type_box.append(key_type_label)
-            key_type_box.append(self.type_dropdown)
-            gen_box.append(key_type_box)
-
-            self.generate_revealer.set_child(gen_box)
-            logger.info("SshCopyIdWindow: Key generation form created successfully")
-        except Exception as e:
-            logger.error(f"SshCopyIdWindow: Failed to create key generation form: {e}")
-            raise
-
-        # Pack into PreferencesGroup
-        logger.info("SshCopyIdWindow: Packing UI elements")
-        try:
-            # Row 1: Existing
-            existing_row = Adw.ActionRow()
-            existing_row.add_prefix(self.radio_existing)
-            existing_row.add_suffix(existing_box)
-            group.add(existing_row)
-
-            # Row 2: Generate
-            generate_row = Adw.ActionRow()
-            generate_row.add_prefix(self.radio_generate)
-            group.add(generate_row)
-            # Embedded generator UI under row 2
-            group.add(self.generate_revealer)
-
-            content.append(group)
-
-            # ---------- Passphrase toggle ----------
-            logger.info("SshCopyIdWindow: Creating passphrase toggle")
-            try:
-                passphrase_group = Adw.PreferencesGroup(title="")
-                
-                self.row_pass_toggle = Adw.SwitchRow()
-                self.row_pass_toggle.set_title("Encrypt with passphrase")
-                self.row_pass_toggle.set_activatable(True)  # Make the entire row clickable
-                
-                passphrase_group.add(self.row_pass_toggle)
-                
-                # Passphrase entries (outside revealer)
-                self.pass1 = Gtk.PasswordEntry()
-                self.pass1.set_property("placeholder-text", "Passphrase")
-                self.pass1.set_can_focus(True)  # Make it focusable for tab navigation
-                self.pass2 = Gtk.PasswordEntry()
-                self.pass2.set_property("placeholder-text", "Confirm passphrase")
-                self.pass2.set_can_focus(True)  # Make it focusable for tab navigation
-                
-                # Create a box for passphrase entries
-                self.pass_box = Gtk.Box(orientation=Gtk.Orientation.VERTICAL, spacing=6)
-                self.pass_box.set_margin_start(12)
-                self.pass_box.set_margin_end(12)
-                self.pass_box.set_margin_top(6)
-                self.pass_box.set_margin_bottom(6)
-                self.pass_box.append(self.pass1)
-                self.pass_box.append(self.pass2)
-                self.pass_box.set_visible(False)
-                
-                passphrase_group.add(self.pass_box)
-                
-                content.append(passphrase_group)
-                logger.info("SshCopyIdWindow: Passphrase toggle created successfully")
-            except Exception as e:
-                logger.error(f"SshCopyIdWindow: Failed to create passphrase toggle: {e}")
-                raise
-
-            # ---------- Force key transfer toggle ----------
-            logger.info("SshCopyIdWindow: Creating force key transfer toggle")
-            try:
-                force_group = Adw.PreferencesGroup(title="")
-                
-                self.force_toggle = Adw.SwitchRow()
-                self.force_toggle.set_title("Force key transfer")
-                self.force_toggle.set_subtitle("Overwrite existing keys on the server")
-                self.force_toggle.set_active(True)  # Default to enabled
-                force_group.add(self.force_toggle)
-                
-                content.append(force_group)
-                logger.info("SshCopyIdWindow: Force key transfer toggle created successfully")
-            except Exception as e:
-                logger.error(f"SshCopyIdWindow: Failed to create force toggle: {e}")
-                raise
-
-            # Radio change behavior
-            self.radio_existing.connect("toggled", self._on_mode_toggled)
-            self.radio_generate.connect("toggled", self._on_mode_toggled)
-            
-            # Set initial state (since "Copy existing key" is selected by default)
-            self.row_pass_toggle.set_sensitive(False)
-            self.pass_box.set_sensitive(False)
-            
-            # Key type change behavior
-            self.type_dropdown.connect("notify::selected", self._on_key_type_changed)
-            
-            # Passphrase toggle behavior
-            self.row_pass_toggle.connect("notify::active", self._on_pass_toggle)
-
-            logger.info("SshCopyIdWindow: UI elements packed successfully")
-        except Exception as e:
-            logger.error(f"SshCopyIdWindow: Failed to pack UI elements: {e}")
-            raise
-
-        logger.info("SshCopyIdWindow: Window construction completed, presenting")
-        self.present()
-
-    # ---------- Helpers ----------
-
-    def _on_mode_toggled(self, *_):
-        # Reveal generator only when "Generate new key" is selected
-        generate_active = self.radio_generate.get_active()
-        logger.info(f"SshCopyIdWindow: Mode toggled, generate active: {generate_active}")
-        self.generate_revealer.set_reveal_child(generate_active)
-        
-        # Enable/disable passphrase section based on mode
-        self.row_pass_toggle.set_sensitive(generate_active)
-        self.pass_box.set_sensitive(generate_active)
-        
-        # If switching to "Copy existing key", turn off passphrase and hide fields
-        if not generate_active:
-            self.row_pass_toggle.set_active(False)
-            self.pass_box.set_visible(False)
-    
-    def _on_key_type_changed(self, *_):
-        # Update key name placeholder when key type changes
-        type_selection = self.type_dropdown.get_selected()
-        if type_selection == 1:  # RSA
-            self.row_key_name.set_text("id_rsa")
-        else:  # ed25519
-            self.row_key_name.set_text("id_ed25519")
-    
-
-    
-    def _on_pass_toggle(self, *_):
-        # Show/hide passphrase entries when toggle changes
-        self.pass_box.set_visible(self.row_pass_toggle.get_active())
-
-    def _reload_existing_keys(self):
-        logger.info("SshCopyIdWindow: Reloading existing keys")
-        logger.debug("SshCopyIdWindow: Calling key_manager.discover_keys()")
-        try:
-            keys = self._km.discover_keys()
-            logger.info(f"SshCopyIdWindow: Discovered {len(keys)} keys")
-            logger.debug(f"SshCopyIdWindow: Key discovery returned {len(keys)} keys")
-            
-            # Log details of each discovered key
-            for i, key in enumerate(keys):
-                logger.debug(f"SshCopyIdWindow: Key {i+1}: private_path='{key.private_path}', "
-                           f"public_path='{key.public_path}', exists={os.path.exists(key.private_path)}")
-            
-            names = [os.path.basename(k.private_path) for k in keys] or ["No keys found"]
-            logger.debug(f"SshCopyIdWindow: Key names for dropdown: {names}")
-            
-            dd = Gtk.DropDown.new_from_strings(names)
-            if keys:
-                dd.set_selected(0)
-                logger.debug(f"SshCopyIdWindow: Selected first key in dropdown")
-            
-            self.dropdown_existing.set_model(dd.get_model())
-            self.dropdown_existing.set_selected(dd.get_selected())
-            # keep a cached list to resolve on OK
-            self._existing_keys_cache = keys
-            logger.info(f"SshCopyIdWindow: Dropdown populated with {len(names)} items")
-            logger.debug(f"SshCopyIdWindow: Cached {len(keys)} keys for later use")
-        except Exception as e:
-            logger.error(f"SshCopyIdWindow: Failed to load existing keys: {e}")
-            logger.debug(f"SshCopyIdWindow: Exception details: {type(e).__name__}: {str(e)}")
-            self._existing_keys_cache = []
-            dd = Gtk.DropDown.new_from_strings(["Error loading keys"])
-            self.dropdown_existing.set_model(dd.get_model())
-            self.dropdown_existing.set_selected(0)
-
-    def _info(self, title, body):
-        try:
-            md = Adw.MessageDialog(transient_for=self, modal=True, heading=title, body=body)
-            md.add_response("ok", "OK")
-            md.set_default_response("ok")
-            md.set_close_response("ok")
-            md.present()
-        except Exception:
-            pass
-
-    def _error(self, title, body, detail=""):
-        try:
-            text = body + (f"\n\n{detail}" if detail else "")
-            md = Adw.MessageDialog(transient_for=self, modal=True, heading=title, body=text)
-            md.add_response("close", "Close")
-            md.set_default_response("close")
-            md.set_close_response("close")
-            md.present()
-        except Exception:
-            logger.error("%s: %s | %s", title, body, detail)
-
-    def _on_close_clicked(self, *_):
-        logger.info("SshCopyIdWindow: Close button clicked")
-        self.close()
-
-    # ---------- OK (main action) ----------
-    def _on_ok_clicked(self, *_):
-        logger.info("SshCopyIdWindow: OK button clicked")
-        logger.debug("SshCopyIdWindow: Starting main action processing")
-        
-        # Log current UI state
-        existing_active = self.radio_existing.get_active()
-        generate_active = self.radio_generate.get_active()
-        logger.debug(f"SshCopyIdWindow: UI state - existing_active={existing_active}, generate_active={generate_active}")
-        
-        try:
-            if self.radio_existing.get_active():
-                logger.info("SshCopyIdWindow: Copying existing key")
-                logger.debug("SshCopyIdWindow: Calling _do_copy_existing()")
-                self._do_copy_existing()
-            else:
-                logger.info("SshCopyIdWindow: Generating new key and copying")
-                logger.debug("SshCopyIdWindow: Calling _do_generate_and_copy()")
-                self._do_generate_and_copy()
-        except Exception as e:
-            logger.error(f"SshCopyIdWindow: Operation failed: {e}")
-            logger.debug(f"SshCopyIdWindow: Exception details: {type(e).__name__}: {str(e)}")
-            self._error("Operation failed", "Could not start the requested action.", str(e))
-
-    # ---------- Mode: existing ----------
-    def _do_copy_existing(self):
-        logger.info("SshCopyIdWindow: Starting copy existing key operation")
-        logger.debug("SshCopyIdWindow: Processing existing key selection")
-        
-        try:
-            keys = getattr(self, "_existing_keys_cache", []) or []
-            logger.info(f"SshCopyIdWindow: Found {len(keys)} cached keys")
-            logger.debug(f"SshCopyIdWindow: Cached keys list length: {len(keys)}")
-            
-            if not keys:
-                logger.debug("SshCopyIdWindow: No cached keys available")
-                raise RuntimeError("No keys available in ~/.ssh")
-            
-            idx = self.dropdown_existing.get_selected()
-            logger.info(f"SshCopyIdWindow: Selected key index: {idx}")
-            logger.debug(f"SshCopyIdWindow: Dropdown selection index: {idx}")
-            
-            if idx < 0 or idx >= len(keys):
-                logger.debug(f"SshCopyIdWindow: Invalid index {idx} for keys list of length {len(keys)}")
-                raise RuntimeError("Please select a key to copy")
-            
-            ssh_key = keys[idx]
-            logger.info(f"SshCopyIdWindow: Selected key: {ssh_key.private_path}")
-            logger.debug(f"SshCopyIdWindow: Selected key details - private_path='{ssh_key.private_path}', "
-                       f"public_path='{ssh_key.public_path}', exists={os.path.exists(ssh_key.private_path)}")
-            
-            # Launch your existing terminal ssh-copy-id flow
-            logger.debug("SshCopyIdWindow: Calling _show_ssh_copy_id_terminal_using_main_widget()")
-            force_enabled = self.force_toggle.get_active()
-            logger.debug(f"SshCopyIdWindow: Force option enabled: {force_enabled}")
-            self._parent._show_ssh_copy_id_terminal_using_main_widget(self._conn, ssh_key, force_enabled)
-            logger.debug("SshCopyIdWindow: Terminal window launched, closing dialog")
-            self.close()
-        except Exception as e:
-            logger.error(f"SshCopyIdWindow: Copy existing failed: {e}")
-            logger.debug(f"SshCopyIdWindow: Exception details: {type(e).__name__}: {str(e)}")
-            self._error("Copy failed", "Could not copy the selected key to the server.", str(e))
-
-    # ---------- Mode: generate ----------
-    def _do_generate_and_copy(self):
-        logger.info("SshCopyIdWindow: Starting generate and copy operation")
-        logger.debug("SshCopyIdWindow: Processing key generation request")
-        
-        try:
-            key_name = (self.row_key_name.get_text() or "").strip()
-            logger.info(f"SshCopyIdWindow: Key name: '{key_name}'")
-            logger.debug(f"SshCopyIdWindow: Raw key name from UI: '{self.row_key_name.get_text()}'")
-            
-            if not key_name:
-                logger.debug("SshCopyIdWindow: Empty key name provided")
-                raise ValueError("Enter a key file name (e.g. id_ed25519)")
-            if "/" in key_name or key_name.startswith("."):
-                logger.debug(f"SshCopyIdWindow: Invalid key name '{key_name}' - contains '/' or starts with '.'")
-                raise ValueError("Key file name must not contain '/' or start with '.'")
-
-            # Key type
-            type_selection = self.type_dropdown.get_selected()
-            kt = "ed25519" if type_selection == 0 else "rsa"
-            logger.info(f"SshCopyIdWindow: Key type: {kt}")
-            logger.debug(f"SshCopyIdWindow: Type selection index: {type_selection}, resolved to: {kt}")
-
-            passphrase = None
-            passphrase_enabled = self.row_pass_toggle.get_active()
-            logger.debug(f"SshCopyIdWindow: Passphrase toggle state: {passphrase_enabled}")
-            
-            if passphrase_enabled:
-                p1 = self.pass1.get_text() or ""
-                p2 = self.pass2.get_text() or ""
-                logger.debug(f"SshCopyIdWindow: Passphrase lengths - p1: {len(p1)}, p2: {len(p2)}")
-                if p1 != p2:
-                    logger.debug("SshCopyIdWindow: Passphrases do not match")
-                    raise ValueError("Passphrases do not match")
-                passphrase = p1
-                logger.info("SshCopyIdWindow: Passphrase enabled")
-                logger.debug("SshCopyIdWindow: Passphrase validation successful")
-
-            logger.info(f"SshCopyIdWindow: Calling key_manager.generate_key with name='{key_name}', type='{kt}'")
-            logger.debug(f"SshCopyIdWindow: Key generation parameters - name='{key_name}', type='{kt}', "
-                       f"size={3072 if kt == 'rsa' else 0}, passphrase={'<set>' if passphrase else 'None'}")
-            
-            new_key = self._km.generate_key(
-                key_name=key_name,
-                key_type=kt,
-                key_size=3072 if kt == "rsa" else 0,
-                comment=None,
-                passphrase=passphrase,
-            )
-            
-            if not new_key:
-                logger.debug("SshCopyIdWindow: Key generation returned None")
-                raise RuntimeError("Key generation failed. See logs for details.")
-
-            logger.info(f"SshCopyIdWindow: Key generated successfully: {new_key.private_path}")
-            logger.debug(f"SshCopyIdWindow: Generated key details - private_path='{new_key.private_path}', "
-                       f"public_path='{new_key.public_path}'")
-            
-            # Ensure the key files are properly written and accessible
-            import time
-            logger.debug("SshCopyIdWindow: Waiting 0.5s for files to be written")
-            time.sleep(0.5)  # Small delay to ensure files are written
-            
-            # Verify the key files exist and are accessible
-            private_exists = os.path.exists(new_key.private_path)
-            public_exists = os.path.exists(new_key.public_path)
-            logger.debug(f"SshCopyIdWindow: File existence check - private: {private_exists}, public: {public_exists}")
-            
-            if not private_exists:
-                logger.debug(f"SshCopyIdWindow: Private key file missing: {new_key.private_path}")
-                raise RuntimeError(f"Private key file not found: {new_key.private_path}")
-            if not public_exists:
-                logger.debug(f"SshCopyIdWindow: Public key file missing: {new_key.public_path}")
-                raise RuntimeError(f"Public key file not found: {new_key.public_path}")
-            
-            logger.info(f"SshCopyIdWindow: Key files verified, starting ssh-copy-id")
-            logger.debug("SshCopyIdWindow: All key files verified successfully")
-            
-            # Run your terminal ssh-copy-id flow
-            logger.debug("SshCopyIdWindow: Calling _show_ssh_copy_id_terminal_using_main_widget()")
-            force_enabled = self.force_toggle.get_active()
-            logger.debug(f"SshCopyIdWindow: Force option enabled: {force_enabled}")
-            self._parent._show_ssh_copy_id_terminal_using_main_widget(self._conn, new_key, force_enabled)
-            logger.debug("SshCopyIdWindow: Terminal window launched, closing dialog")
-            self.close()
-
-        except Exception as e:
-            logger.error(f"SshCopyIdWindow: Generate and copy failed: {e}")
-            logger.debug(f"SshCopyIdWindow: Exception details: {type(e).__name__}: {str(e)}")
-            self._error("Generate & Copy failed",
-                        "Could not generate a new key and copy it to the server.",
-                        str(e))
 
 
 # ============================================================================
@@ -2036,1173 +657,6 @@ class WelcomePage(Gtk.Box):
             shortcuts_box.append(shortcut_box)
         
         self.append(shortcuts_box)
-
-class MonospaceFontDialog(Adw.Window):
-    def __init__(self, parent=None, current_font="Monospace 12"):
-        super().__init__()
-        
-        self.set_title("Select Terminal Font")
-        self.set_default_size(500, 600)
-        self.set_transient_for(parent)
-        self.set_modal(True)
-        
-        # Store callback
-        self.callback = None
-        
-        # Parse current font
-        self.current_font_desc = Pango.FontDescription.from_string(current_font)
-        
-        # Create main content
-        self.setup_ui()
-        self.populate_fonts()
-        
-    def setup_ui(self):
-        # Main box
-        main_box = Gtk.Box(orientation=Gtk.Orientation.VERTICAL, spacing=12)
-        main_box.set_margin_top(12)
-        main_box.set_margin_bottom(12)
-        main_box.set_margin_start(12)
-        main_box.set_margin_end(12)
-        
-        # Header
-        header = Adw.HeaderBar()
-        header.set_show_start_title_buttons(False)
-        header.set_show_end_title_buttons(False)
-        
-        # Cancel button
-        cancel_btn = Gtk.Button(label="Cancel")
-        cancel_btn.connect("clicked", self.on_cancel)
-        header.pack_start(cancel_btn)
-        
-        # Select button
-        select_btn = Gtk.Button(label="Select")
-        select_btn.add_css_class("suggested-action")
-        select_btn.connect("clicked", self.on_select)
-        header.pack_end(select_btn)
-        
-        # Create search entry
-        self.search_entry = Gtk.SearchEntry()
-        self.search_entry.set_placeholder_text("Search fonts...")
-        self.search_entry.connect("search-changed", self.on_search_changed)
-        
-        # Create font list
-        self.font_model = Gtk.ListStore(str, str, object)  # display_name, family, font_desc
-        self.font_filter = Gtk.TreeModelFilter(child_model=self.font_model)
-        self.font_filter.set_visible_func(self.filter_fonts)
-        
-        self.font_view = Gtk.TreeView(model=self.font_filter)
-        self.font_view.set_headers_visible(False)
-        
-        # Font name column
-        name_renderer = Gtk.CellRendererText()
-        name_column = Gtk.TreeViewColumn("Font", name_renderer, text=0)
-        self.font_view.append_column(name_column)
-        
-        # Selection handling
-        selection = self.font_view.get_selection()
-        selection.connect("changed", self.on_selection_changed)
-        
-        # Scrolled window for font list
-        scrolled = Gtk.ScrolledWindow()
-        scrolled.set_policy(Gtk.PolicyType.NEVER, Gtk.PolicyType.AUTOMATIC)
-        scrolled.set_child(self.font_view)
-        scrolled.set_vexpand(True)
-        
-        # Size selection
-        size_box = Gtk.Box(orientation=Gtk.Orientation.HORIZONTAL, spacing=6)
-        size_label = Gtk.Label(label="Size:")
-        size_label.set_halign(Gtk.Align.START)
-        
-        self.size_spin = Gtk.SpinButton.new_with_range(6, 72, 1)
-        self.size_spin.set_value(self.current_font_desc.get_size() / Pango.SCALE)
-        self.size_spin.connect("value-changed", self.on_size_changed)
-        
-        size_box.append(size_label)
-        size_box.append(self.size_spin)
-        
-        # Preview text
-        preview_frame = Gtk.Frame()
-        preview_frame.set_label("Preview")
-        
-        self.preview_label = Gtk.Label()
-        self.preview_label.set_text("The quick brown fox jumps over the lazy dog\n0123456789 !@#$%^&*()_+-=[]{}|;:,.<>?")
-        self.preview_label.set_margin_top(12)
-        self.preview_label.set_margin_bottom(12)
-        self.preview_label.set_margin_start(12)
-        self.preview_label.set_margin_end(12)
-        self.preview_label.set_selectable(True)
-        preview_frame.set_child(self.preview_label)
-        
-        # Add everything to main box
-        main_box.append(header)
-        main_box.append(self.search_entry)
-        main_box.append(scrolled)
-        main_box.append(size_box)
-        main_box.append(preview_frame)
-        
-        self.set_content(main_box)
-        
-    def populate_fonts(self):
-        # Get font map
-        fontmap = PangoFT2.FontMap.new()
-        families = fontmap.list_families()
-        
-        monospace_families = []
-        for family in families:
-            if family.is_monospace():
-                family_name = family.get_name()
-                
-                # Get available faces for this family
-                faces = family.list_faces()
-                for face in faces:
-                    face_name = face.get_face_name()
-                    
-                    # Create font description
-                    desc = Pango.FontDescription()
-                    desc.set_family(family_name)
-                    desc.set_size(12 * Pango.SCALE)
-                    
-                    # Set style if not regular
-                    if face_name.lower() != "regular":
-                        if "bold" in face_name.lower():
-                            desc.set_weight(Pango.Weight.BOLD)
-                        if "italic" in face_name.lower() or "oblique" in face_name.lower():
-                            desc.set_style(Pango.Style.ITALIC)
-                    
-                    # Display name
-                    if face_name.lower() == "regular":
-                        display_name = family_name
-                    else:
-                        display_name = f"{family_name} {face_name}"
-                    
-                    monospace_families.append((display_name, family_name, desc))
-        
-        # Sort by family name
-        monospace_families.sort(key=lambda x: x[0].lower())
-        
-        # Add to model
-        for display_name, family_name, desc in monospace_families:
-            self.font_model.append([display_name, family_name, desc])
-            
-        # Select current font if possible
-        self.select_current_font()
-        
-    def select_current_font(self):
-        current_family = self.current_font_desc.get_family()
-        if not current_family:
-            return
-            
-        iter = self.font_model.get_iter_first()
-        while iter:
-            family = self.font_model.get_value(iter, 1)
-            if family.lower() == current_family.lower():
-                # Convert to filter iter
-                filter_iter = self.font_filter.convert_child_iter_to_iter(iter)
-                if filter_iter[1]:  # Check if conversion was successful
-                    selection = self.font_view.get_selection()
-                    selection.select_iter(filter_iter[1])
-                    
-                    # Scroll to selection
-                    path = self.font_filter.get_path(filter_iter[1])
-                    self.font_view.scroll_to_cell(path, None, False, 0.0, 0.0)
-                break
-            iter = self.font_model.iter_next(iter)
-    
-    def filter_fonts(self, model, iter, data):
-        search_text = self.search_entry.get_text().lower()
-        if not search_text:
-            return True
-            
-        font_name = model.get_value(iter, 0).lower()
-        return search_text in font_name
-    
-    def on_search_changed(self, entry):
-        self.font_filter.refilter()
-    
-    def on_selection_changed(self, selection):
-        model, iter = selection.get_selected()
-        if iter:
-            font_desc = model.get_value(iter, 2).copy()
-            font_desc.set_size(int(self.size_spin.get_value()) * Pango.SCALE)
-            self.update_preview(font_desc)
-    
-    def on_size_changed(self, spin):
-        selection = self.font_view.get_selection()
-        model, iter = selection.get_selected()
-        if iter:
-            font_desc = model.get_value(iter, 2).copy()
-            font_desc.set_size(int(spin.get_value()) * Pango.SCALE)
-            self.update_preview(font_desc)
-    
-    def update_preview(self, font_desc):
-        # Remove previous CSS provider if it exists
-        if hasattr(self, '_css_provider'):
-            context = self.preview_label.get_style_context()
-            context.remove_provider(self._css_provider)
-        
-        # Create CSS for the font
-        css = f"""
-        .preview-font {{
-            font-family: "{font_desc.get_family()}";
-            font-size: {font_desc.get_size() / Pango.SCALE}pt;
-            font-weight: normal;
-            font-style: normal;
-        """
-        
-        if font_desc.get_weight() == Pango.Weight.BOLD:
-            css += "font-weight: bold;"
-        if font_desc.get_style() == Pango.Style.ITALIC:
-            css += "font-style: italic;"
-            
-        css += "}"
-        
-        # Apply CSS
-        self._css_provider = Gtk.CssProvider()
-        self._css_provider.load_from_data(css.encode())
-        
-        context = self.preview_label.get_style_context()
-        context.add_provider(self._css_provider, Gtk.STYLE_PROVIDER_PRIORITY_APPLICATION)
-        
-        # Ensure the class is added (but only once)
-        if not context.has_class("preview-font"):
-            context.add_class("preview-font")
-    
-    def on_cancel(self, button):
-        self.close()
-    
-    def on_select(self, button):
-        selection = self.font_view.get_selection()
-        model, iter = selection.get_selected()
-        if iter and self.callback:
-            font_desc = model.get_value(iter, 2).copy()
-            font_desc.set_size(int(self.size_spin.get_value()) * Pango.SCALE)
-            font_string = font_desc.to_string()
-            self.callback(font_string)
-        self.close()
-    
-    def set_callback(self, callback):
-        """Set callback function that receives the selected font string"""
-        self.callback = callback
-
-
-class PreferencesWindow(Adw.PreferencesWindow):
-    """Preferences dialog window"""
-    
-    def __init__(self, parent_window, config):
-        super().__init__()
-        self.set_transient_for(parent_window)
-        self.set_modal(True)
-        self.config = config
-        
-        # Set window properties
-        self.set_title("Preferences")
-        self.set_default_size(600, 500)
-        
-        # Initialize the preferences UI
-        self.setup_preferences()
-
-        # Save on close to persist advanced SSH settings
-        self.connect('close-request', self.on_close_request)
-    
-    def setup_preferences(self):
-        """Set up preferences UI with current values"""
-        try:
-            # Create Terminal preferences page
-            terminal_page = Adw.PreferencesPage()
-            terminal_page.set_title("Terminal")
-            terminal_page.set_icon_name("utilities-terminal-symbolic")
-            
-            # Terminal appearance group
-            appearance_group = Adw.PreferencesGroup()
-            appearance_group.set_title("Appearance")
-            
-            # Font selection row
-            self.font_row = Adw.ActionRow()
-            self.font_row.set_title("Font")
-            current_font = self.config.get_setting('terminal-font', 'Monospace 12')
-            self.font_row.set_subtitle(current_font)
-            
-            font_button = Gtk.Button()
-            font_button.set_label("Choose")
-            font_button.connect('clicked', self.on_font_button_clicked)
-            self.font_row.add_suffix(font_button)
-            
-            appearance_group.add(self.font_row)
-            
-            # Terminal color scheme
-            self.color_scheme_row = Adw.ComboRow()
-            self.color_scheme_row.set_title("Color Scheme")
-            self.color_scheme_row.set_subtitle("Terminal color theme")
-            
-            color_schemes = Gtk.StringList()
-            color_schemes.append("Default")
-            color_schemes.append("Solarized Dark")
-            color_schemes.append("Solarized Light")
-            color_schemes.append("Monokai")
-            color_schemes.append("Dracula")
-            color_schemes.append("Nord")
-            color_schemes.append("Gruvbox Dark")
-            color_schemes.append("One Dark")
-            color_schemes.append("Tomorrow Night")
-            color_schemes.append("Material Dark")
-            self.color_scheme_row.set_model(color_schemes)
-            
-            # Set current color scheme from config
-            current_scheme_key = self.config.get_setting('terminal.theme', 'default')
-            
-            # Get the display name for the current scheme key
-            theme_mapping = self.get_theme_name_mapping()
-            reverse_mapping = {v: k for k, v in theme_mapping.items()}
-            current_scheme_display = reverse_mapping.get(current_scheme_key, 'Default')
-            
-            # Find the index of the current scheme in the dropdown
-            scheme_names = [
-                "Default", "Solarized Dark", "Solarized Light",
-                "Monokai", "Dracula", "Nord",
-                "Gruvbox Dark", "One Dark", "Tomorrow Night", "Material Dark"
-            ]
-            try:
-                current_index = scheme_names.index(current_scheme_display)
-                self.color_scheme_row.set_selected(current_index)
-            except ValueError:
-                # If the saved scheme isn't found, default to the first option
-                self.color_scheme_row.set_selected(0)
-                # Also update the config to use the default value
-                self.config.set_setting('terminal.theme', 'default')
-            
-            self.color_scheme_row.connect('notify::selected', self.on_color_scheme_changed)
-            
-            appearance_group.add(self.color_scheme_row)
-            
-            # Color scheme preview
-            preview_group = Adw.PreferencesGroup()
-            preview_group.set_title("Preview")
-            
-            # Create preview terminal widget
-            self.color_preview_terminal = Gtk.DrawingArea()
-            self.color_preview_terminal.set_draw_func(self.draw_color_preview)
-            self.color_preview_terminal.set_size_request(400, 120)
-            self.color_preview_terminal.add_css_class("terminal-preview")
-            
-            # Add some margin around the preview
-            preview_box = Gtk.Box(orientation=Gtk.Orientation.VERTICAL)
-            preview_box.set_margin_top(6)
-            preview_box.set_margin_bottom(6)
-            preview_box.set_margin_start(12)
-            preview_box.set_margin_end(12)
-            preview_box.append(self.color_preview_terminal)
-            
-            preview_group.add(preview_box)
-            appearance_group.add(preview_group)
-            terminal_page.add(appearance_group)
-            
-            # Preferred Terminal group (only show when not in Flatpak)
-            if not is_running_in_flatpak():
-                terminal_choice_group = Adw.PreferencesGroup()
-                terminal_choice_group.set_title("Preferred Terminal")
-                
-                # Radio buttons for terminal choice
-                self.builtin_terminal_radio = Gtk.CheckButton(label="Use built-in terminal")
-                self.builtin_terminal_radio.set_can_focus(True)
-                self.external_terminal_radio = Gtk.CheckButton(label="Use other terminal")
-                self.external_terminal_radio.set_can_focus(True)
-                
-                # Make them behave like radio buttons
-                self.external_terminal_radio.set_group(self.builtin_terminal_radio)
-                
-                # Set current preference
-                use_external = self.config.get_setting('use-external-terminal', False)
-                if use_external:
-                    self.external_terminal_radio.set_active(True)
-                else:
-                    self.builtin_terminal_radio.set_active(True)
-                
-                # Connect radio button changes
-                self.builtin_terminal_radio.connect('toggled', self.on_terminal_choice_changed)
-                self.external_terminal_radio.connect('toggled', self.on_terminal_choice_changed)
-                
-                # Add radio buttons to group
-                terminal_choice_group.add(self.builtin_terminal_radio)
-                terminal_choice_group.add(self.external_terminal_radio)
-                
-                # External terminal dropdown and custom path
-                self.external_terminal_box = Gtk.Box(orientation=Gtk.Orientation.VERTICAL, spacing=6)
-                self.external_terminal_box.set_margin_start(24)
-                self.external_terminal_box.set_margin_top(6)
-                self.external_terminal_box.set_margin_bottom(6)
-                
-                # Terminal dropdown
-                self.terminal_dropdown = Gtk.DropDown()
-                self.terminal_dropdown.set_can_focus(True)
-                
-                # Populate dropdown with available terminals
-                self._populate_terminal_dropdown()
-                
-                # Set current selection
-                current_terminal = self.config.get_setting('external-terminal', 'gnome-terminal')
-                self._set_terminal_dropdown_selection(current_terminal)
-                
-                # Show/hide custom path entry based on current selection
-                if hasattr(self, 'custom_terminal_box'):
-                    if current_terminal == 'custom':
-                        self.custom_terminal_box.set_visible(True)
-                    else:
-                        self.custom_terminal_box.set_visible(False)
-                
-                # Connect dropdown changes
-                self.terminal_dropdown.connect('notify::selected', self.on_terminal_dropdown_changed)
-                
-                # Custom path entry (initially hidden)
-                self.custom_terminal_box = Gtk.Box(orientation=Gtk.Orientation.HORIZONTAL, spacing=6)
-                self.custom_terminal_entry = Gtk.Entry()
-                self.custom_terminal_entry.set_placeholder_text("/usr/bin/gnome-terminal")
-                self.custom_terminal_entry.set_can_focus(True)
-                
-                # Set current custom path if any
-                custom_path = self.config.get_setting('custom-terminal-path', '')
-                if custom_path:
-                    self.custom_terminal_entry.set_text(custom_path)
-                
-                # Connect custom path changes
-                self.custom_terminal_entry.connect('changed', self.on_custom_terminal_path_changed)
-                
-                self.custom_terminal_box.append(self.custom_terminal_entry)
-                
-                # Initially hide custom path entry
-                self.custom_terminal_box.set_visible(False)
-                
-                # Add dropdown and custom path to box
-                self.external_terminal_box.append(self.terminal_dropdown)
-                self.external_terminal_box.append(self.custom_terminal_box)
-                
-                # Initial sensitivity will be set by radio button state
-                
-                # Add to group
-                terminal_choice_group.add(self.external_terminal_box)
-                
-                # Set initial sensitivity based on radio button state
-                self.external_terminal_box.set_sensitive(self.external_terminal_radio.get_active())
-                
-                terminal_page.add(terminal_choice_group)
-            
-            # Create Interface preferences page
-            interface_page = Adw.PreferencesPage()
-            interface_page.set_title("Interface")
-            interface_page.set_icon_name("applications-graphics-symbolic")
-            
-            # App startup behavior
-            startup_group = Adw.PreferencesGroup()
-            startup_group.set_title("App Startup")
-            
-            # Radio buttons for startup behavior
-            self.terminal_startup_radio = Gtk.CheckButton(label="Show Terminal")
-            self.terminal_startup_radio.set_can_focus(True)
-            self.welcome_startup_radio = Gtk.CheckButton(label="Show Welcome Page")
-            self.welcome_startup_radio.set_can_focus(True)
-            
-            # Make them behave like radio buttons
-            self.welcome_startup_radio.set_group(self.terminal_startup_radio)
-            
-            # Set current preference (default to terminal)
-            startup_behavior = self.config.get_setting('app-startup-behavior', 'terminal')
-            if startup_behavior == 'welcome':
-                self.welcome_startup_radio.set_active(True)
-            else:
-                self.terminal_startup_radio.set_active(True)
-            
-            # Connect radio button changes
-            self.terminal_startup_radio.connect('toggled', self.on_startup_behavior_changed)
-            self.welcome_startup_radio.connect('toggled', self.on_startup_behavior_changed)
-            
-            # Add radio buttons to group
-            startup_group.add(self.terminal_startup_radio)
-            startup_group.add(self.welcome_startup_radio)
-            
-            interface_page.add(startup_group)
-            
-            # Appearance group
-            interface_appearance_group = Adw.PreferencesGroup()
-            interface_appearance_group.set_title("Appearance")
-            
-            # Theme selection
-            self.theme_row = Adw.ComboRow()
-            self.theme_row.set_title("Application Theme")
-            self.theme_row.set_subtitle("Choose light, dark, or follow system theme")
-            
-            themes = Gtk.StringList()
-            themes.append("Follow System")
-            themes.append("Light")
-            themes.append("Dark")
-            self.theme_row.set_model(themes)
-            
-            # Load saved theme preference
-            saved_theme = self.config.get_setting('app-theme', 'default')
-            theme_mapping = {'default': 0, 'light': 1, 'dark': 2}
-            self.theme_row.set_selected(theme_mapping.get(saved_theme, 0))
-            
-            self.theme_row.connect('notify::selected', self.on_theme_changed)
-            
-            interface_appearance_group.add(self.theme_row)
-            interface_page.add(interface_appearance_group)
-            
-            # Window group
-            window_group = Adw.PreferencesGroup()
-            window_group.set_title("Window")
-            
-            # Remember window size switch
-            remember_size_switch = Adw.SwitchRow()
-            remember_size_switch.set_title("Remember Window Size")
-            remember_size_switch.set_subtitle("Restore window size on startup")
-            remember_size_switch.set_active(True)
-            
-            # Auto focus terminal switch
-            auto_focus_switch = Adw.SwitchRow()
-            auto_focus_switch.set_title("Auto Focus Terminal")
-            auto_focus_switch.set_subtitle("Focus terminal when connecting")
-            auto_focus_switch.set_active(True)
-            
-            window_group.add(remember_size_switch)
-            window_group.add(auto_focus_switch)
-            interface_page.add(window_group)
-
-            # Advanced SSH settings
-            advanced_page = Adw.PreferencesPage()
-            advanced_page.set_title("Advanced")
-            advanced_page.set_icon_name("applications-system-symbolic")
-
-            advanced_group = Adw.PreferencesGroup()
-            advanced_group.set_title("SSH Settings")
-            # Use custom options toggle
-            self.apply_advanced_row = Adw.SwitchRow()
-            self.apply_advanced_row.set_title("Use custom connection options")
-            self.apply_advanced_row.set_subtitle("Enable and edit the options below")
-            self.apply_advanced_row.set_active(bool(self.config.get_setting('ssh.apply_advanced', False)))
-            advanced_group.add(self.apply_advanced_row)
-
-
-            # Connect timeout
-            self.connect_timeout_row = Adw.SpinRow.new_with_range(1, 120, 1)
-            self.connect_timeout_row.set_title("Connect Timeout (s)")
-            self.connect_timeout_row.set_value(self.config.get_setting('ssh.connection_timeout', 10))
-            advanced_group.add(self.connect_timeout_row)
-
-            # Connection attempts
-            self.connection_attempts_row = Adw.SpinRow.new_with_range(1, 10, 1)
-            self.connection_attempts_row.set_title("Connection Attempts")
-            self.connection_attempts_row.set_value(self.config.get_setting('ssh.connection_attempts', 1))
-            advanced_group.add(self.connection_attempts_row)
-
-            # Keepalive interval
-            self.keepalive_interval_row = Adw.SpinRow.new_with_range(0, 300, 5)
-            self.keepalive_interval_row.set_title("ServerAlive Interval (s)")
-            self.keepalive_interval_row.set_value(self.config.get_setting('ssh.keepalive_interval', 30))
-            advanced_group.add(self.keepalive_interval_row)
-
-            # Keepalive count max
-            self.keepalive_count_row = Adw.SpinRow.new_with_range(1, 10, 1)
-            self.keepalive_count_row.set_title("ServerAlive CountMax")
-            self.keepalive_count_row.set_value(self.config.get_setting('ssh.keepalive_count_max', 3))
-            advanced_group.add(self.keepalive_count_row)
-
-            # Strict host key checking
-            self.strict_host_row = Adw.ComboRow()
-            self.strict_host_row.set_title("StrictHostKeyChecking")
-            strict_model = Gtk.StringList()
-            for item in ["accept-new", "yes", "no", "ask"]:
-                strict_model.append(item)
-            self.strict_host_row.set_model(strict_model)
-            # Map current value
-            current_strict = str(self.config.get_setting('ssh.strict_host_key_checking', 'accept-new'))
-            try:
-                idx = ["accept-new", "yes", "no", "ask"].index(current_strict)
-            except ValueError:
-                idx = 0
-            self.strict_host_row.set_selected(idx)
-            advanced_group.add(self.strict_host_row)
-
-            # BatchMode (non-interactive)
-            self.batch_mode_row = Adw.SwitchRow()
-            self.batch_mode_row.set_title("BatchMode (disable prompts)")
-            self.batch_mode_row.set_active(bool(self.config.get_setting('ssh.batch_mode', True)))
-            advanced_group.add(self.batch_mode_row)
-
-            # Compression
-            self.compression_row = Adw.SwitchRow()
-            self.compression_row.set_title("Enable Compression (-C)")
-            self.compression_row.set_active(bool(self.config.get_setting('ssh.compression', True)))
-            advanced_group.add(self.compression_row)
-
-            # SSH verbosity (-v levels)
-            self.verbosity_row = Adw.SpinRow.new_with_range(0, 3, 1)
-            self.verbosity_row.set_title("SSH Verbosity (-v)")
-            self.verbosity_row.set_value(int(self.config.get_setting('ssh.verbosity', 0)))
-            advanced_group.add(self.verbosity_row)
-
-            # Debug logging toggle
-            self.debug_enabled_row = Adw.SwitchRow()
-            self.debug_enabled_row.set_title("Enable SSH Debug Logging")
-            self.debug_enabled_row.set_active(bool(self.config.get_setting('ssh.debug_enabled', False)))
-            advanced_group.add(self.debug_enabled_row)
-
-            # Confirm before disconnecting
-            self.confirm_disconnect_switch = Adw.SwitchRow()
-            self.confirm_disconnect_switch.set_title("Confirm before disconnecting")
-            self.confirm_disconnect_switch.set_subtitle("Show a confirmation dialog when disconnecting from a host")
-            self.confirm_disconnect_switch.set_active(
-                self.config.get_setting('confirm-disconnect', True)
-            )
-            self.confirm_disconnect_switch.connect('notify::active', self.on_confirm_disconnect_changed)
-            advanced_group.add(self.confirm_disconnect_switch)
-
-            # Reset button
-            # Add spacing before reset button
-            advanced_group.add(Gtk.Separator(orientation=Gtk.Orientation.HORIZONTAL))
-            
-            # Use Adw.ActionRow for proper spacing and layout
-            reset_row = Adw.ActionRow()
-            reset_row.set_title("Reset Advanced SSH Settings")
-            reset_row.set_subtitle("Restore all advanced SSH settings to their default values")
-            
-            reset_btn = Gtk.Button.new_with_label("Reset")
-            reset_btn.add_css_class('destructive-action')
-            reset_btn.connect('clicked', self.on_reset_advanced_ssh)
-            reset_row.add_suffix(reset_btn)
-            
-            advanced_group.add(reset_row)
-
-            # Disable/enable advanced controls based on toggle
-            def _sync_advanced_sensitivity(row=None, *_):
-                enabled = bool(self.apply_advanced_row.get_active())
-                for w in [self.connect_timeout_row, self.connection_attempts_row,
-                          self.keepalive_interval_row, self.keepalive_count_row,
-                          self.strict_host_row, self.batch_mode_row,
-                          self.compression_row, self.verbosity_row,
-                          self.debug_enabled_row]:
-                    try:
-                        w.set_sensitive(enabled)
-                    except Exception:
-                        pass
-            _sync_advanced_sensitivity()
-            self.apply_advanced_row.connect('notify::active', _sync_advanced_sensitivity)
-
-            advanced_page.add(advanced_group)
-
-            # Add pages to the preferences window
-            self.add(interface_page)
-            self.add(terminal_page)
-            self.add(advanced_page)
-            
-            logger.info("Preferences window initialized")
-        except Exception as e:
-            logger.error(f"Failed to setup preferences: {e}")
-
-    def on_close_request(self, *args):
-        """Persist settings when the preferences window closes"""
-        try:
-            self.save_advanced_ssh_settings()
-            # Ensure preferences are flushed to disk
-            if hasattr(self.config, 'save_json_config'):
-                self.config.save_json_config()
-        except Exception:
-            pass
-        return False  # allow close
-    
-    def on_font_button_clicked(self, button):
-        """Handle font button click"""
-        logger.info("Font button clicked")
-        
-        # Get current font from config
-        current_font = self.config.get_setting('terminal-font', 'Monospace 12')
-        
-        # Create custom monospace font dialog
-        font_dialog = MonospaceFontDialog(parent=self, current_font=current_font)
-        
-        def on_font_selected(font_string):
-            self.font_row.set_subtitle(font_string)
-            logger.info(f"Font selected: {font_string}")
-            
-            # Save to config
-            self.config.set_setting('terminal-font', font_string)
-            
-            # Apply to all active terminals
-            self.apply_font_to_terminals(font_string)
-        
-        font_dialog.set_callback(on_font_selected)
-        font_dialog.present()
-    
-    def apply_font_to_terminals(self, font_string):
-        """Apply font to all active terminal widgets"""
-        try:
-            parent_window = self.get_transient_for()
-            if parent_window and hasattr(parent_window, 'connection_to_terminals'):
-                font_desc = Pango.FontDescription.from_string(font_string)
-                count = 0
-                for terms in parent_window.connection_to_terminals.values():
-                    for terminal in terms:
-                        if hasattr(terminal, 'vte'):
-                            terminal.vte.set_font(font_desc)
-                            count += 1
-                logger.info(f"Applied font {font_string} to {count} terminals")
-        except Exception as e:
-            logger.error(f"Failed to apply font to terminals: {e}")
-    
-    def on_theme_changed(self, combo_row, param):
-        """Handle theme selection change"""
-        selected = combo_row.get_selected()
-        theme_names = ["Follow System", "Light", "Dark"]
-        selected_theme = theme_names[selected] if selected < len(theme_names) else "Follow System"
-        
-        logger.info(f"Theme changed to: {selected_theme}")
-        
-        # Apply theme immediately
-        style_manager = Adw.StyleManager.get_default()
-        
-        if selected == 0:  # Follow System
-            style_manager.set_color_scheme(Adw.ColorScheme.DEFAULT)
-            self.config.set_setting('app-theme', 'default')
-        elif selected == 1:  # Light
-            style_manager.set_color_scheme(Adw.ColorScheme.FORCE_LIGHT)
-            self.config.set_setting('app-theme', 'light')
-        elif selected == 2:  # Dark
-            style_manager.set_color_scheme(Adw.ColorScheme.FORCE_DARK)
-            self.config.set_setting('app-theme', 'dark')
-
-    def save_advanced_ssh_settings(self):
-        """Persist advanced SSH settings from the preferences UI"""
-        try:
-            if hasattr(self, 'apply_advanced_row'):
-                self.config.set_setting('ssh.apply_advanced', bool(self.apply_advanced_row.get_active()))
-            if hasattr(self, 'connect_timeout_row'):
-                self.config.set_setting('ssh.connection_timeout', int(self.connect_timeout_row.get_value()))
-            if hasattr(self, 'connection_attempts_row'):
-                self.config.set_setting('ssh.connection_attempts', int(self.connection_attempts_row.get_value()))
-            if hasattr(self, 'keepalive_interval_row'):
-                self.config.set_setting('ssh.keepalive_interval', int(self.keepalive_interval_row.get_value()))
-            if hasattr(self, 'keepalive_count_row'):
-                self.config.set_setting('ssh.keepalive_count_max', int(self.keepalive_count_row.get_value()))
-            if hasattr(self, 'strict_host_row'):
-                options = ["accept-new", "yes", "no", "ask"]
-                idx = self.strict_host_row.get_selected()
-                value = options[idx] if 0 <= idx < len(options) else 'accept-new'
-                self.config.set_setting('ssh.strict_host_key_checking', value)
-            if hasattr(self, 'batch_mode_row'):
-                self.config.set_setting('ssh.batch_mode', bool(self.batch_mode_row.get_active()))
-            if hasattr(self, 'compression_row'):
-                self.config.set_setting('ssh.compression', bool(self.compression_row.get_active()))
-            if hasattr(self, 'verbosity_row'):
-                self.config.set_setting('ssh.verbosity', int(self.verbosity_row.get_value()))
-            if hasattr(self, 'debug_enabled_row'):
-                self.config.set_setting('ssh.debug_enabled', bool(self.debug_enabled_row.get_active()))
-        except Exception as e:
-            logger.error(f"Failed to save advanced SSH settings: {e}")
-
-    def on_reset_advanced_ssh(self, *args):
-        """Reset only advanced SSH keys to defaults and update UI."""
-        try:
-            defaults = self.config.get_default_config().get('ssh', {})
-            # Persist defaults and disable apply
-            self.config.set_setting('ssh.apply_advanced', False)
-            for key in ['connection_timeout', 'connection_attempts', 'keepalive_interval', 'keepalive_count_max', 'compression', 'auto_add_host_keys', 'verbosity', 'debug_enabled']:
-                self.config.set_setting(f'ssh.{key}', defaults.get(key))
-            # Update UI
-            if hasattr(self, 'apply_advanced_row'):
-                self.apply_advanced_row.set_active(False)
-            if hasattr(self, 'connect_timeout_row'):
-                self.connect_timeout_row.set_value(int(defaults.get('connection_timeout', 30)))
-            if hasattr(self, 'connection_attempts_row'):
-                self.connection_attempts_row.set_value(int(defaults.get('connection_attempts', 1)))
-            if hasattr(self, 'keepalive_interval_row'):
-                self.keepalive_interval_row.set_value(int(defaults.get('keepalive_interval', 60)))
-            if hasattr(self, 'keepalive_count_row'):
-                self.keepalive_count_row.set_value(int(defaults.get('keepalive_count_max', 3)))
-            if hasattr(self, 'strict_host_row'):
-                try:
-                    self.strict_host_row.set_selected(["accept-new", "yes", "no", "ask"].index('accept-new'))
-                except ValueError:
-                    self.strict_host_row.set_selected(0)
-            if hasattr(self, 'batch_mode_row'):
-                self.batch_mode_row.set_active(False)
-            if hasattr(self, 'compression_row'):
-                self.compression_row.set_active(bool(defaults.get('compression', True)))
-            if hasattr(self, 'verbosity_row'):
-                self.verbosity_row.set_value(int(defaults.get('verbosity', 0)))
-            if hasattr(self, 'debug_enabled_row'):
-                self.debug_enabled_row.set_active(bool(defaults.get('debug_enabled', False)))
-        except Exception as e:
-            logger.error(f"Failed to reset advanced SSH settings: {e}")
-    
-    def get_theme_name_mapping(self):
-        """Get mapping between display names and config keys"""
-        return {
-            "Default": "default",
-            "Solarized Dark": "solarized_dark", 
-            "Solarized Light": "solarized_light",
-            "Monokai": "monokai",
-            "Dracula": "dracula",
-            "Nord": "nord",
-            "Gruvbox Dark": "gruvbox_dark",
-            "One Dark": "one_dark",
-            "Tomorrow Night": "tomorrow_night",
-            "Material Dark": "material_dark",
-        }
-    
-    def get_reverse_theme_mapping(self):
-        """Get mapping from config keys to display names"""
-        mapping = self.get_theme_name_mapping()
-        return {v: k for k, v in mapping.items()}
-    
-    def on_color_scheme_changed(self, combo_row, param):
-        """Handle terminal color scheme change"""
-        selected = combo_row.get_selected()
-        scheme_names = [
-            "Default", "Solarized Dark", "Solarized Light",
-            "Monokai", "Dracula", "Nord",
-            "Gruvbox Dark", "One Dark", "Tomorrow Night", "Material Dark"
-        ]
-        selected_scheme = scheme_names[selected] if selected < len(scheme_names) else "Default"
-        
-        logger.info(f"Terminal color scheme changed to: {selected_scheme}")
-        
-        # Convert display name to config key
-        theme_mapping = self.get_theme_name_mapping()
-        config_key = theme_mapping.get(selected_scheme, "default")
-        
-        # Save to config using the consistent key
-        self.config.set_setting('terminal.theme', config_key)
-        
-        # Apply to all active terminals
-        self.apply_color_scheme_to_terminals(config_key)
-        
-        # Refresh the color preview
-        if hasattr(self, 'color_preview_terminal'):
-            self.color_preview_terminal.queue_draw()
-    
-    def draw_color_preview(self, drawing_area, cr, width, height):
-        """Draw a preview of the selected color scheme"""
-        # Get current color scheme
-        current_scheme_key = self.config.get_setting('terminal.theme', 'default')
-        
-        # Get color scheme colors
-        colors = self.get_color_scheme_colors(current_scheme_key)
-        
-        # Draw background
-        bg_color = colors.get('background', '#000000')
-        r, g, b, a = self.hex_to_rgba(bg_color)
-        cr.set_source_rgba(r, g, b, a)
-        cr.paint()
-        
-        # Draw terminal-like content
-        cr.set_font_size(10)
-        
-        # Draw prompt line
-        prompt_color = colors.get('foreground', '#ffffff')
-        r, g, b, a = self.hex_to_rgba(prompt_color)
-        cr.set_source_rgba(r, g, b, a)
-        cr.move_to(10, 25)
-        cr.show_text("user@host:~$ ")
-        
-        # Draw command
-        command_color = colors.get('foreground', '#ffffff')
-        r, g, b, a = self.hex_to_rgba(command_color)
-        cr.set_source_rgba(r, g, b, a)
-        cr.move_to(120, 25)
-        cr.show_text("ls -la")
-        
-        # Draw output lines
-        output_color = colors.get('foreground', '#ffffff')
-        r, g, b, a = self.hex_to_rgba(output_color)
-        cr.set_source_rgba(r, g, b, a)
-        
-        # Directory line
-        cr.move_to(10, 45)
-        cr.show_text("drwxr-xr-x  2 user user 4096 Jan 15 10:30 .")
-        
-        # File line with different color
-        file_color = colors.get('blue', '#0088ff')
-        r, g, b, a = self.hex_to_rgba(file_color)
-        cr.set_source_rgba(r, g, b, a)
-        cr.move_to(10, 65)
-        cr.show_text("-rw-r--r--  1 user user  1234 Jan 15 10:25 file.txt")
-        
-        # Executable file
-        exec_color = colors.get('green', '#00ff00')
-        r, g, b, a = self.hex_to_rgba(exec_color)
-        cr.set_source_rgba(r, g, b, a)
-        cr.move_to(10, 85)
-        cr.show_text("-rwxr-xr-x  1 user user 5678 Jan 15 10:20 script.sh")
-        
-        # Prompt line 2
-        prompt_color = colors.get('foreground', '#ffffff')
-        r, g, b, a = self.hex_to_rgba(prompt_color)
-        cr.set_source_rgba(r, g, b, a)
-        cr.move_to(10, 105)
-        cr.show_text("user@host:~$ ")
-        
-    def get_color_scheme_colors(self, scheme_key):
-        """Get colors for a specific color scheme"""
-        schemes = {
-            'default': {
-                'background': '#000000',
-                'foreground': '#ffffff',
-                'blue': '#0088ff',
-                'green': '#00ff00',
-                'red': '#ff0000',
-                'yellow': '#ffff00',
-                'magenta': '#ff00ff',
-                'cyan': '#00ffff'
-            },
-            'solarized_dark': {
-                'background': '#002b36',
-                'foreground': '#839496',
-                'blue': '#268bd2',
-                'green': '#859900',
-                'red': '#dc322f',
-                'yellow': '#b58900',
-                'magenta': '#d33682',
-                'cyan': '#2aa198'
-            },
-            'solarized_light': {
-                'background': '#fdf6e3',
-                'foreground': '#657b83',
-                'blue': '#268bd2',
-                'green': '#859900',
-                'red': '#dc322f',
-                'yellow': '#b58900',
-                'magenta': '#d33682',
-                'cyan': '#2aa198'
-            },
-            'monokai': {
-                'background': '#272822',
-                'foreground': '#f8f8f2',
-                'blue': '#66d9ef',
-                'green': '#a6e22e',
-                'red': '#f92672',
-                'yellow': '#e6db74',
-                'magenta': '#fd5ff0',
-                'cyan': '#a1efe4'
-            },
-            'dracula': {
-                'background': '#282a36',
-                'foreground': '#f8f8f2',
-                'blue': '#6272a4',
-                'green': '#50fa7b',
-                'red': '#ff5555',
-                'yellow': '#f1fa8c',
-                'magenta': '#bd93f9',
-                'cyan': '#8be9fd'
-            },
-            'nord': {
-                'background': '#2e3440',
-                'foreground': '#eceff4',
-                'blue': '#5e81ac',
-                'green': '#a3be8c',
-                'red': '#bf616a',
-                'yellow': '#ebcb8b',
-                'magenta': '#b48ead',
-                'cyan': '#88c0d0'
-            },
-            'gruvbox_dark': {
-                'background': '#282828',
-                'foreground': '#ebdbb2',
-                'blue': '#83a598',
-                'green': '#b8bb26',
-                'red': '#fb4934',
-                'yellow': '#fabd2f',
-                'magenta': '#d3869b',
-                'cyan': '#8ec07c'
-            },
-            'one_dark': {
-                'background': '#282c34',
-                'foreground': '#abb2bf',
-                'blue': '#61afef',
-                'green': '#98c379',
-                'red': '#e06c75',
-                'yellow': '#e5c07b',
-                'magenta': '#c678dd',
-                'cyan': '#56b6c2'
-            },
-            'tomorrow_night': {
-                'background': '#1d1f21',
-                'foreground': '#c5c8c6',
-                'blue': '#81a2be',
-                'green': '#b5bd68',
-                'red': '#cc6666',
-                'yellow': '#f0c674',
-                'magenta': '#b294bb',
-                'cyan': '#8abeb7'
-            },
-            'material_dark': {
-                'background': '#263238',
-                'foreground': '#eeffff',
-                'blue': '#82aaff',
-                'green': '#c3e88d',
-                'red': '#f07178',
-                'yellow': '#ffcb6b',
-                'magenta': '#c792ea',
-                'cyan': '#89ddff'
-            }
-        }
-        return schemes.get(scheme_key, schemes['default'])
-    
-    def hex_to_rgba(self, hex_color):
-        """Convert hex color to RGBA values (0-1 range)"""
-        hex_color = hex_color.lstrip('#')
-        r = int(hex_color[0:2], 16) / 255.0
-        g = int(hex_color[2:4], 16) / 255.0
-        b = int(hex_color[4:6], 16) / 255.0
-        return (r, g, b, 1.0)
-        
-    def on_confirm_disconnect_changed(self, switch, *args):
-        """Handle confirm disconnect setting change"""
-        confirm = switch.get_active()
-        logger.info(f"Confirm before disconnect setting changed to: {confirm}")
-        self.config.set_setting('confirm-disconnect', confirm)
-    
-    def on_startup_behavior_changed(self, radio_button, *args):
-        """Handle startup behavior radio button change"""
-        show_terminal = self.terminal_startup_radio.get_active()
-        behavior = 'terminal' if show_terminal else 'welcome'
-        logger.info(f"App startup behavior changed to: {behavior}")
-        self.config.set_setting('app-startup-behavior', behavior)
-    
-    def on_terminal_choice_changed(self, radio_button, *args):
-        """Handle terminal choice radio button change"""
-        use_external = self.external_terminal_radio.get_active()
-        logger.info(f"Terminal choice changed to: {'external' if use_external else 'built-in'}")
-        
-        # Enable/disable external terminal options
-        self.external_terminal_box.set_sensitive(use_external)
-        
-        # Save preference
-        self.config.set_setting('use-external-terminal', use_external)
-    
-    def on_terminal_dropdown_changed(self, dropdown, *args):
-        """Handle terminal dropdown selection change"""
-        selected = dropdown.get_selected()
-        if selected is None or selected < 0:
-            return
-        
-        # Get the selected terminal from the model
-        model = dropdown.get_model()
-        if model and selected < model.get_n_items():
-            terminal_name = model.get_string(selected)
-            logger.info(f"Terminal dropdown selection changed to: {terminal_name}")
-            
-            # Show/hide custom path entry based on selection
-            if hasattr(self, 'custom_terminal_box'):
-                if terminal_name == "Custom":
-                    self.custom_terminal_box.set_visible(True)
-                else:
-                    self.custom_terminal_box.set_visible(False)
-            
-            # Save the selected terminal
-            if terminal_name != "Custom":
-                self.config.set_setting('external-terminal', terminal_name)
-    
-    def on_custom_terminal_path_changed(self, entry, *args):
-        """Handle custom terminal path entry change"""
-        custom_path = entry.get_text().strip()
-        logger.info(f"Custom terminal path changed to: {custom_path}")
-        
-        # Validate the path
-        if custom_path and self._is_valid_unix_path(custom_path):
-            self.config.set_setting('custom-terminal-path', custom_path)
-            self.config.set_setting('external-terminal', 'custom')
-        else:
-            # Clear invalid path
-            self.config.set_setting('custom-terminal-path', '')
-    
-    def _populate_terminal_dropdown(self):
-        """Populate the terminal dropdown with available terminals"""
-        try:
-            # Create string list for dropdown
-            terminals_list = Gtk.StringList()
-            
-            # Add common terminals
-            common_terminals = [
-                'gnome-terminal', 'konsole', 'xfce4-terminal', 'alacritty', 
-                'kitty', 'terminator', 'tilix', 'xterm'
-            ]
-            
-            # Check which terminals are available
-            available_terminals = []
-            for terminal in common_terminals:
-                try:
-                    result = subprocess.run(['which', terminal], capture_output=True, text=True, timeout=2)
-                    if result.returncode == 0:
-                        available_terminals.append(terminal)
-                except Exception:
-                    continue
-            
-            # Add available terminals to dropdown
-            for terminal in available_terminals:
-                terminals_list.append(terminal)
-            
-            # Add "Custom" option
-            terminals_list.append("Custom")
-            
-            # Set the model
-            self.terminal_dropdown.set_model(terminals_list)
-            
-            logger.info(f"Populated terminal dropdown with {len(available_terminals)} available terminals")
-            
-        except Exception as e:
-            logger.error(f"Failed to populate terminal dropdown: {e}")
-    
-    def _set_terminal_dropdown_selection(self, terminal_name):
-        """Set the dropdown selection to the specified terminal"""
-        try:
-            model = self.terminal_dropdown.get_model()
-            if not model:
-                return
-            
-            # Find the terminal in the model
-            for i in range(model.get_n_items()):
-                if model.get_string(i) == terminal_name:
-                    self.terminal_dropdown.set_selected(i)
-                    return
-            
-            # If not found, default to first available terminal
-            if model.get_n_items() > 0:
-                self.terminal_dropdown.set_selected(0)
-                
-        except Exception as e:
-            logger.error(f"Failed to set terminal dropdown selection: {e}")
-    
-    def _is_valid_unix_path(self, path):
-        """Validate if the path is a valid Unix path"""
-        if not path:
-            return False
-        
-        # Check if it starts with / (absolute path)
-        if not path.startswith('/'):
-            return False
-        
-        # Check if it contains only valid characters
-        import re
-        if not re.match(r'^[a-zA-Z0-9/._-]+$', path):
-            return False
-        
-        # Check if the file exists and is executable
-        try:
-            return os.path.isfile(path) and os.access(path, os.X_OK)
-        except Exception:
-            return False
-    
-    def apply_color_scheme_to_terminals(self, scheme_key):
-        """Apply color scheme to all active terminal widgets"""
-        try:
-            parent_window = self.get_transient_for()
-            if parent_window and hasattr(parent_window, 'connection_to_terminals'):
-                count = 0
-                for terms in parent_window.connection_to_terminals.values():
-                    for terminal in terms:
-                        if hasattr(terminal, 'apply_theme'):
-                            terminal.apply_theme(scheme_key)
-                            count += 1
-                logger.info(f"Applied color scheme {scheme_key} to {count} terminals")
-        except Exception as e:
-            logger.error(f"Failed to apply color scheme to terminals: {e}")
 
 class MainWindow(Adw.ApplicationWindow):
     """Main application window"""
